### 2019-04-23 #1405 @whedon
[PRE REVIEW]: insight: Easy Access to Model Information for Various Model Objects
**Submitting author:** @strengejacke (<a href="http://orcid.org/0000-0002-8895-3206">Daniel Lüdecke</a>)
**Repository:** <a href="https://github.com/easystats/insight" target ="_blank">https://github.com/easystats/insight</a>
**Version:** 0.2.1
