### 2019-04-23 #1405 @danielskatz
@strengejacke - I don't see a paper.bib file in your repo - please see https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography for an example of how the paper.md and paper.bib work together
