### 2019-04-23 #1405 @strengejacke
Thanks @danielskatz for starting the pre-review. Actually, the package is more "technical" and does not implement methods proposed in (statistical) papers... That's why we haven't included a bibliography. If it's required, we could at least cite the R-software, if this is OK for you?
