### 2019-04-23 #1405 @strengejacke
Ok, bibliography added.
Suggested reviewers, with no preferred order:

@amoeba
@cschwem2er
@Ohlsen
@jsta
@corybrunson
