### 2019-04-23 #1405 @danielskatz
Ok, it looks like you don't actually have to cite anything, but you do need a paper.bib file.
If you want to cite R, you will need to put a reference to it in the paper.md file - see the example for how to do this.  If not, I guess it's fine, but it's a bit odd to have a paper about research software that has no citations of previous work.  The reviewers can discuss this issue...
(If you do make changes, you can regenerate the PDF by adding a new comment here containing ```@whedon generate pdf```)
