### 2019-04-23 #1405 @strengejacke
I have cited R in the appropriate place, but indeed I forgot a _References_ heading at the end...
