### 2019-04-25 6382461
Small typo fix in fig caption
added "users" after "allow" and before "to" in line 36
