### 2019-04-25 d06b347
fix get_random() for brms and stanreg
