+++ NAMESPACE
@@ -85,6 +85,7 @@ S3method(get_data,BFBayesFactor)
 S3method(get_data,LORgee)
 S3method(get_data,MCMCglmm)
 S3method(get_data,MixMod)
+S3method(get_data,brmsfit)
 S3method(get_data,clm2)
 S3method(get_data,clmm)
 S3method(get_data,data.frame)
@@ -107,6 +108,7 @@ S3method(get_data,plm)
 S3method(get_data,rlmerMod)
 S3method(get_data,rqss)
 S3method(get_data,stanmvreg)
+S3method(get_data,stanreg)
 S3method(get_data,tobit)
 S3method(get_data,vgam)
 S3method(get_data,vglm)
