### 2019-04-25 b9bbed5
fix for rms::rcs() functions in get_data()
