### 2019-04-26 2579dbb
close #85, give more informative warning
