+++ tests/testthat/test-aovlist.R
@@ -1,5 +1,5 @@
 if (require("testthat") && require("insight") && require("stats")) {
-  context("insight, model_info")
+  context("insight, aovlist")
 
   data(npk)
   m1 <- aov(yield ~  N*P*K + Error(block), data = npk)
