+++ NAMESPACE
@@ -142,6 +142,7 @@ S3method(get_parameters,stanreg)
 S3method(get_parameters,vgam)
 S3method(get_parameters,zeroinfl)
 S3method(get_parameters,zerotrunc)
+S3method(get_priors,BFBayesFactor)
 S3method(get_priors,brmsfit)
 S3method(get_priors,stanreg)
 S3method(get_variance,MixMod)
