+++ tests/testthat/test-BayesFactorBF.R
@@ -171,4 +171,21 @@ if (require("testthat") && require("insight") && require("stats") && require("Ba
     expect_null(get_parameters(x))
   })
 
+  test_that("get_priors", {
+    expect_equal(
+      get_priors(x),
+      structure(
+        list(
+          parameters = c("fixed", "random", "continuous"),
+          distribution = c("Cauchy", "Cauchy", "Cauchy"),
+          location = c(0, 0, 0),
+          scale = c(0.5, 1, 0.353553390593274)
+        ),
+        class = "data.frame",
+        row.names = c("fixed", "random", "continuous")
+      ),
+      tolerance = 1e-5
+    )
+  })
+
 }
