+++ NAMESPACE
@@ -360,6 +360,7 @@ importFrom(stats,Gamma)
 importFrom(stats,as.formula)
 importFrom(stats,binomial)
 importFrom(stats,coef)
+importFrom(stats,drop.terms)
 importFrom(stats,family)
 importFrom(stats,formula)
 importFrom(stats,gaussian)
