### 2019-04-30 7f67dc7
fix triple edge case
