+++ tests/testthat/test-lmer.R
@@ -174,17 +174,20 @@ if (require("testthat") && require("insight") && require("lme4")) {
     skip_on_cran()
     skip_on_travis()
 
-    expect_equal(get_variance(m1), list(
-      var.fixed = 908.95336262316459396970,
-      var.random = 1698.23306388298283309268,
-      var.residual = 654.94079585243218843971,
-      var.distribution = 654.94079585243218843971,
-      var.dispersion = 0,
-      var.intercept = c(Subject = 611.89760710463770010392),
-      var.slope = c(Subject.Days = 35.08106944030500073950),
-      cor.slope_intercept = c(Subject = 0.06561803142425107205)
-    ),
-    tolerance = 1e-4)
+    expect_equal(
+      get_variance(m1),
+      list(
+        var.fixed = 908.953362623165,
+        var.random = 1698.23306388298,
+        var.residual = 654.940795852432,
+        var.distribution = 654.940795852432,
+        var.dispersion = 0,
+        var.intercept = c(Subject = 611.897607104638),
+        var.slope = c(Subject.Days = 35.081069440305),
+        cor.slope_intercept = c(Subject = 0.0656180314242511)
+      ),
+      tolerance = 1e-4
+    )
 
     expect_equal(get_variance_fixed(m1), c(var.fixed = 908.95336262316459396970), tolerance = 1e-4)
     expect_equal(get_variance_random(m1), c(var.random = 1698.23306388298283309268), tolerance = 1e-4)
@@ -196,19 +199,7 @@ if (require("testthat") && require("insight") && require("lme4")) {
     expect_equal(get_variance_slope(m1), c(var.slope.Subject.Days = 35.08106944030500073950), toleance = 1e-4)
     expect_equal(get_correlation_slope_intercept(m1), c(cor.slope_intercept.Subject = 0.06561803), toleance = 1e-4)
 
-    expect_equal(get_variance(m2), list(
-      var.fixed = 899.71188516552831515583,
-      var.random = 1412.32769480994716104760,
-      var.residual = 928.81331970697453925823,
-      var.distribution = 928.81331970697453925823,
-      var.dispersion = 0,
-      var.intercept = c(
-        `mysubgrp:mygrp` = 12.61789198780912713005,
-        Subject = 1376.88290392979729404033,
-        mygrp = 22.82689889233859048545
-      )
-    ),
-    tolerance = 1e-4)
+    expect_warning(expect_equal(get_variance(m2), NA))
   })
 
   test_that("find_algorithm", {
