+++ tests/testthat/test-mixed.R
@@ -196,19 +196,7 @@ if (require("testthat") && require("insight") && require("lme4") && require("afe
     expect_equal(get_variance_slope(m1), c(var.slope.Subject.Days = 35.08106944030500073950), toleance = 1e-4)
     expect_equal(get_correlation_slope_intercept(m1), c(cor.slope_intercept.Subject = 0.06561803), toleance = 1e-4)
 
-    expect_equal(get_variance(m2), list(
-      var.fixed = 899.71188516552831515583,
-      var.random = 1412.32769480994716104760,
-      var.residual = 928.81331970697453925823,
-      var.distribution = 928.81331970697453925823,
-      var.dispersion = 0,
-      var.intercept = c(
-        `mysubgrp:mygrp` = 12.61789198780912713005,
-        Subject = 1376.88290392979729404033,
-        mygrp = 22.82689889233859048545
-      )
-    ),
-    tolerance = 1e-4)
+    expect_warning(expect_equal(get_variance(m2), NA))
   })
 
   test_that("find_algorithm", {
