### 2019-05-02 068a37c
support speedglm
