+++ DESCRIPTION
@@ -63,6 +63,7 @@ Suggests:
   robustlmm,
   rstanarm,
   rstudioapi,
+  speedglm,
   splines,
   survey,
   survival,
