+++ NAMESPACE
@@ -23,6 +23,8 @@ S3method(find_algorithm,mixed)
 S3method(find_algorithm,rlmerMod)
 S3method(find_algorithm,rq)
 S3method(find_algorithm,rqss)
+S3method(find_algorithm,speedglm)
+S3method(find_algorithm,speedlm)
 S3method(find_algorithm,stanreg)
 S3method(find_formula,BFBayesFactor)
 S3method(find_formula,LORgee)
@@ -189,6 +191,8 @@ S3method(link_function,plm)
 S3method(link_function,polr)
 S3method(link_function,psm)
 S3method(link_function,rq)
+S3method(link_function,speedglm)
+S3method(link_function,speedlm)
 S3method(link_function,stanmvreg)
 S3method(link_function,survreg)
 S3method(link_function,tobit)
@@ -240,6 +244,8 @@ S3method(link_inverse,plm)
 S3method(link_inverse,polr)
 S3method(link_inverse,psm)
 S3method(link_inverse,rq)
+S3method(link_inverse,speedglm)
+S3method(link_inverse,speedlm)
 S3method(link_inverse,stanmvreg)
 S3method(link_inverse,survreg)
 S3method(link_inverse,tobit)
@@ -293,6 +299,8 @@ S3method(model_info,plm)
 S3method(model_info,polr)
 S3method(model_info,rq)
 S3method(model_info,rqss)
+S3method(model_info,speedglm)
+S3method(model_info,speedlm)
 S3method(model_info,stanmvreg)
 S3method(model_info,survreg)
 S3method(model_info,tobit)

[…]
