+++ R/link_function.R
@@ -60,6 +60,12 @@ link_function.speedglm <- function(x, ...) {
 
 
 #' @export
+link_function.bigglm <- function(x, ...) {
+  stats::family(x)$linkfun
+}
+
+
+#' @export
 link_function.multinom <- function(x, ...) {
   stats::make.link(link = "logit")$linkfun
 }
@@ -263,6 +269,12 @@ link_function.lmRob <- function(x, ...) {
 
 #' @export
 link_function.speedlm <- function(x, ...) {
+  stats::gaussian(link = "identity")$linkfun
+}
+
+
+#' @export
+link_function.biglm <- function(x, ...) {
   stats::gaussian(link = "identity")$linkfun
 }
 
