+++ DESCRIPTION
@@ -26,6 +26,7 @@ Suggests:
   afex,
   BayesFactor,
   betareg,
+  biglm,
   blme,
   brms,
   censReg,
