### 2019-05-03 1feea27
function no longer return "1" or "0", related to https://github.com/easystats/insight/issues/92
