### 2019-05-03 95434d6
fix issue with brms-formula, check may run on CRAN?
