### 2019-05-03 fc83132
run tests also on CRAN
