### 2019-05-03 4a47c39
check if tests can also run on Travis
