### 2019-05-03 069a843
add funs to check for intercept and nullmodel
