+++ NEWS.md
@@ -4,6 +4,11 @@
 
 * `biglm` and `bigglm` (*biglm*), `feis` (*feisr*), `gbm` (*gbm*), `BFBayesFactor` (*BayesFactor*), `psm` (*rms*), `LORgee` (*multgee*), `censReg` (*censReg*), `ols` (*rms*), `speedlm` and `speedglm` (*speedglm*), `svyolr` (*survey*)
 
+## New functions
+
+* `is_nullmodel()` to check if model is a null-model (intercept-only), i.e. if the conditional part of the model has no predictors.
+* `has_intercept()` to check if model has an intercept.
+
 ## Breaking Changes
 
 * Functions like `find_predictors()` or `find_terms()` return `NULL` for null-models (intercept-only models). Use `find_parameters()` to check if your model only has an intercept-parameter (but no predictors).
