+++ NAMESPACE
@@ -370,8 +370,10 @@ export(get_variance_intercept)
 export(get_variance_random)
 export(get_variance_residual)
 export(get_variance_slope)
+export(has_intercept)
 export(is_model)
 export(is_multivariate)
+export(is_nullmodel)
 export(link_function)
 export(link_inverse)
 export(model_info)
