+++ NEWS.md
@@ -21,7 +21,7 @@
 * `get_random()` failed for *brmsfit* and *stanreg* models.
 * `get_parameters()` and `find_parameters()` did not work for `MixMod`-objects _without_ zero-inflation component, when `component = "all"` (the default).
 * `find_formula()` did not work for `plm`-models without instrumental variables.
-* `find_formula()` returned random effects as conditional part of the formula for null-models (only intercept in fixed parts) (#87)-
+* `find_formula()` returned random effects as conditional part of the formula for null-models (only intercept in fixed parts) (#87).
 * Fixed issue with invalid notation of instrumental-variables formula in `felm`-models for R-devel on Linux.
 * Fixed issue with `get_data()` for *gee* models, where incomplete cases were not removed from the data.
 * Fixed potential issue with `get_data()` for null-models (only intercept in fixed parts) from models of class `glmmTMB`, `brmsfit`, `MixMod` and `rstanarm` (#91).
