### 2019-05-04 fd93f9a
add covr to suggests
