+++ DESCRIPTION
@@ -32,6 +32,7 @@ Suggests:
   censReg,
   crch,
   coxme,
+  covr,
   estimatr,
   feisr,
   gam,
