### 2019-05-05 1a95bdf
tests work on travis, maybe not on CRAN, though
