+++ tests/testthat/test-rstanarm.R
@@ -1,7 +1,6 @@
-# .runThisTest <- Sys.getenv("RunAllinsightTests") == "yes"
-# if (.runThisTest && Sys.getenv("USER") != "travis") {
+.runThisTest <- Sys.getenv("RunAllinsightTests") == "yes"
+# if (.runThisTest) {
 
-#if (Sys.getenv("USER") != "travis") {
   if (suppressWarnings(
     require("testthat") &&
       require("insight") &&
@@ -150,4 +149,5 @@
       ))
     })
   }
+
 # }
