### 2019-05-05 5118d41
fix issue with get_priors
