### 2019-05-06 53abf3f
add flatten to find_parms()
