+++ R/find_parameters.R
@@ -32,112 +32,163 @@
 #' find_parameters(m)
 #' @importFrom stats coef
 #' @export
-find_parameters <- function(x, ...) {
+find_parameters <- function(x, flatten = FALSE, ...) {
   UseMethod("find_parameters")
 }
 
 
 #' @export
-find_parameters.default <- function(x, ...) {
+find_parameters.default <- function(x, flatten = FALSE, ...) {
   if (inherits(x, "list") && obj_has_name(x, "gam")) {
     x <- x$gam
     class(x) <- c(class(x), c("glm", "lm"))
-    return(find_parameters.gam(x))
+    pars <- find_parameters.gam(x)
+  } else {
+    pars <- tryCatch({
+      list(conditional = names(stats::coef(x)))
+    },
+    error = function(x) {
+      NULL
+    }
+    )
   }
 
-  tryCatch({
-    list(conditional = names(stats::coef(x)))
-  },
-  error = function(x) {
-    NULL
+  if (flatten) {
+    unique(unlist(pars))
+  } else {
+    pars
   }
-  )
 }
 
 
 #' @export
-find_parameters.data.frame <- function(x, ...) {
+find_parameters.data.frame <- function(x, flatten = FALSE, ...) {
   stop("A data frame is no valid object for this function")
 }
 
 
 #' @export
-find_parameters.gamlss <- function(x, ...) {
+find_parameters.gamlss <- function(x, flatten = FALSE, ...) {
   pars <- list(
     conditional = names(stats::coef(x)),
     sigma = names(stats::coef(x, what = "sigma")),
     nu = names(stats::coef(x, what = "nu")),
     tau = names(stats::coef(x, what = "tau"))
   )
-  compact_list(pars)
+
+  pars <- compact_list(pars)
+
+  if (flatten) {
+    unique(unlist(pars))
+  } else {
+    pars
+  }
 }
 
 
 #' @export
-find_parameters.gam <- function(x, ...) {
+find_parameters.gam <- function(x, flatten = FALSE, ...) {
   pars <- list(conditional = names(stats::coef(x)))
   st <- summary(x)$s.table
 
   pars$conditional <- pars$conditional[.grep_non_smoothers(pars$conditional)]
   pars$smooth_terms <- row.names(st)
 
-  compact_list(pars)
+  pars <- compact_list(pars)
+
+  if (flatten) {
+    unique(unlist(pars))
+  } else {
+    pars
+  }
 }
 
 
 #' @export
-find_parameters.gbm <- function(x, ...) {
+find_parameters.gbm <- function(x, flatten = FALSE, ...) {
   s <- summary(x, plotit = FALSE)
-  list(conditional = as.character(s$var))
+  pars <- list(conditional = as.character(s$var))
+
+  if (flatten) {
+    unique(unlist(pars))
+  } else {
+    pars
+  }
 }
 
 
 #' @export
-find_parameters.Gam <- function(x, ...) {
+find_parameters.Gam <- function(x, flatten = FALSE, ...) {
   pars <- names(stats::coef(x))
 
-  compact_list(list(
+  l <- compact_list(list(
     conditional = pars[.grep_non_smoothers(pars)],
     smooth_terms = pars[.grep_smoothers(pars)]
   ))
+
+  if (flatten) {
+    unique(unlist(l))
+  } else {
+    l
+  }
 }
 
 
 #' @export
-find_parameters.vgam <- function(x, ...) {
+find_parameters.vgam <- function(x, flatten = FALSE, ...) {
   pars <- names(stats::coef(x))
 
-  compact_list(list(
[…]
