### 2019-05-06 f149c70
fix get_parmameters
