### 2019-05-06 9b82834
prepare r2 for mixed-zi
