+++ man/dot-get_variance_poisson_family.Rd
@@ -1,0 +1,12 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/compute_variances.R
+\name{.get_variance_poisson_family}
+\alias{.get_variance_poisson_family}
+\title{Get distributional variance for poisson-family}
+\usage{
+.get_variance_poisson_family(x, mu, faminfo)
+}
+\description{
+Get distributional variance for poisson-family
+}
+\keyword{internal}
