### 2019-05-06 3e91a62
mu from null-model
