### 2019-05-06 588bc47
get_variance accounts for zi-part
