+++ NEWS.md
@@ -15,6 +15,7 @@
 
 ## Changes to functions
 
+* `get_variance()` now computes the full variance for mixed models with zero-inflation component.
 * `get_priors()` now returns the default-prior that was defined for all parameters of a class, if certain parameters have no specific prior.
 * `find_parameters()` gets a `flatten`-argument, to either return results as list or as simple vector.
 
