+++ NAMESPACE
@@ -396,6 +396,7 @@ importFrom(stats,na.omit)
 importFrom(stats,nobs)
 importFrom(stats,plogis)
 importFrom(stats,poisson)
+importFrom(stats,predict)
 importFrom(stats,reformulate)
 importFrom(stats,terms)
 importFrom(stats,update)
