### 2019-05-06 37ac96b
get_variance works for zi-models
