### 2019-05-07 066f31e
variance for zip is simpler
