### 2019-05-07 e795bc5
update get_variance
