+++ R/compute_variances.R
@@ -336,7 +336,7 @@
 #'
 #' @importFrom stats family
 #' @keywords internal
-.get_variance_dist <- function(x, null.fixef, faminfo, sig, name, verbose = TRUE) {
+.get_variance_dist <- function(x, null_model, faminfo, sig, name, verbose = TRUE) {
   if (!requireNamespace("lme4", quietly = TRUE)) {
     stop("Package `lme4` needs to be installed to compute variances for mixed models.", call. = FALSE)
   }
@@ -346,6 +346,7 @@
 
   # in general want log(1+var(x)/mu^2)
 
+  null.fixef <- unname(.collapse_cond(lme4::fixef(null_model)))
   mu <- exp(null.fixef)
 
   if (is.na(mu)) {
@@ -364,7 +365,7 @@
     vv <- switch(
       faminfo$family,
       `zero-inflated poisson` = ,
-      poisson                 = .get_variance_poisson_family(x, mu, faminfo),
+      poisson                 = .get_variance_poisson_family(x, null_model, mu, faminfo),
       `hurdle poisson`    = ,
       truncated_poisson   = stats::family(x)$variance(sig),
       tweedie             = .get_variance_tweedie_family(x, mu, sig),
@@ -373,7 +374,7 @@
       `negative binomial` = ,
       genpois             = ,
       nbinom1             = ,
-      nbinom2             = .get_variance_nbinom_family(x, mu, sig, faminfo),
+      nbinom2             = .get_variance_nbinom_family(x, null_model, mu, sig, faminfo),
       .get_variance_default(x, mu, verbose)
     )
 
@@ -395,9 +396,9 @@
 #' Get distributional variance for poisson-family
 #'
 #' @keywords internal
-.get_variance_poisson_family <- function(x, mu, faminfo) {
+.get_variance_poisson_family <- function(x, null_model, mu, faminfo) {
   if (faminfo$is_zeroinf) {
-    .get_variance_zip(x, mu, faminfo, family_var = mu)
+    .get_variance_zip(x, null_model, faminfo, family_var = mu)
   } else {
     if (inherits(x, "MixMod")) {
       return(mu)
@@ -435,10 +436,10 @@
 #' Get distributional variance for nbinom-family
 #'
 #' @keywords internal
-.get_variance_nbinom_family <- function(x, mu, sig, faminfo) {
+.get_variance_nbinom_family <- function(x, null_model, mu, sig, faminfo) {
   if (faminfo$is_zeroinf) {
     if (missing(sig)) sig <- 0
-    .get_variance_zinb(x, mu, sig, faminfo, family_var = mu * (1 + sig))
+    .get_variance_zinb(x, null_model, sig, faminfo, family_var = mu * (1 + sig))
   } else {
     if (inherits(x, "MixMod")) {
       if (missing(sig))
@@ -454,12 +455,12 @@
 
 #' @importFrom stats plogis family predict
 #' @keywords internal
-.get_variance_zinb <- function(model, mu, sig, faminfo, family_var) {
+.get_variance_zinb <- function(model, null_model, sig, faminfo, family_var) {
   if (inherits(model, "glmmTMB")) {
     v <- stats::family(model)$variance
     p <- stats::predict(model, type = "zprob")  ## z-i probability
     ## mean of conditional distribution
-    # mu <- stats::predict(model, type = "conditional")
+    mu <- stats::predict(null_model, type = "conditional")
     betad <- model$fit$par["betad"]
     k <- switch(
       faminfo$family,
@@ -471,7 +472,8 @@
   } else if (inherits(model, "MixMod")) {
     v <- family_var
     ## z-i probability
[…]
