+++ R/compute_variances.R
@@ -2,9 +2,12 @@
 #' @keywords internal
 .compute_variances <- function(x, component, name_fun = NULL, name_full = NULL, verbose = TRUE) {
 
-  ## Code taken from GitGub-Repo of package glmmTMB
-  ## Author: Ben Bolker, who used an
-  ## cleaned-up/adapted version of Jon Lefcheck's code from SEMfit
+  ## Original code taken from GitGub-Repo of package glmmTMB
+  ## Author: Ben Bolker, who used an cleaned-up/adapted
+  ## version of Jon Lefcheck's code from SEMfit
+
+  ## Major revisions and adaption to more complex models and other packages
+  ## by Daniel Lüdecke
 
   faminfo <- model_info(x)
 
@@ -114,6 +117,7 @@
 
 
 
+
 #' store essential information on coefficients, model matrix and so on
 #' as list, since we need these information throughout the functions to
 #' calculate the variance components...
@@ -196,7 +200,8 @@
 
 
 
-#' helper-function, telling user if model is supported or not
+
+#' helper-function, telling user if family / distribution is supported or not
 #'
 #' @keywords internal
 .badlink <- function(link, family, verbose = TRUE) {
@@ -208,8 +213,10 @@
 
 
 
+
 #' glmmTMB returns a list of model information, one for conditional
-#' and one for zero-inflated part, so here we "unlist" it
+#' and one for zero-inflated part, so here we "unlist" it, returning
+#' only the conditional part.
 #'
 #' @keywords internal
 .collapse_cond <- function(x) {
@@ -222,6 +229,7 @@
 
 
 
+
 #' Get fixed effects variance
 #'
 #' @importFrom stats var
@@ -229,6 +237,8 @@
 .compute_variance_fixed <- function(vals) {
   with(vals, stats::var(as.vector(beta %*% t(X))))
 }
+
+
 
 
 
@@ -262,6 +272,8 @@
 }
 
 
+
+
 #' Calculate Distribution-specific variance (Nakagawa et al. 2017)
 #'
 #' @keywords internal
@@ -289,20 +301,20 @@
     } else if (faminfo$is_count) {
       dist.variance <- switch(
         faminfo$link_function,
-        log = .get_variance_dist(x, faminfo, sig, name = name, verbose = verbose),
+        log = .variance_distributional(x, faminfo, sig, name = name, verbose = verbose),
         sqrt = 0.25,
         .badlink(faminfo$link_function, faminfo$family, verbose = verbose)
       )
     } else if (faminfo$family == "beta") {
       dist.variance <- switch(
         faminfo$link_function,
-        logit = .get_variance_dist(x, faminfo, sig, name = name, verbose = verbose),
+        logit = .variance_distributional(x, faminfo, sig, name = name, verbose = verbose),
         .badlink(faminfo$link_function, faminfo$family, verbose = verbose)
       )
     } else if (faminfo$is_tweedie) {
       dist.variance <- switch(
         faminfo$link_function,
-        log = .get_variance_dist(x, faminfo, sig, name = name, verbose = verbose),
+        log = .variance_distributional(x, faminfo, sig, name = name, verbose = verbose),
         .badlink(faminfo$link_function, faminfo$family, verbose = verbose)
       )
     }
@@ -310,6 +322,7 @@
 
   dist.variance
 }
+
 
 
 
@@ -330,13 +343,14 @@
 
 
 
+
 #' This is the core-function to calculate the distribution-specific variance
 #' Nakagawa et al. 2017 propose three different methods, here we only rely
 #' on the lognormal-approximation.
 #'
[…]
