+++ man/dot-variance_distributional.Rd
@@ -1,12 +1,12 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/compute_variances.R
-\name{.get_variance_dist}
-\alias{.get_variance_dist}
+\name{.variance_distributional}
+\alias{.variance_distributional}
 \title{This is the core-function to calculate the distribution-specific variance
 Nakagawa et al. 2017 propose three different methods, here we only rely
 on the lognormal-approximation.}
 \usage{
-.get_variance_dist(x, faminfo, sig, name, verbose = TRUE)
+.variance_distributional(x, faminfo, sig, name, verbose = TRUE)
 }
 \description{
 This is the core-function to calculate the distribution-specific variance
