+++ man/dot-variance_family_beta.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/compute_variances.R
-\name{.get_variance_beta_family}
-\alias{.get_variance_beta_family}
+\name{.variance_family_beta}
+\alias{.variance_family_beta}
 \title{Get distributional variance for beta-family}
 \usage{
-.get_variance_beta_family(x, mu, phi)
+.variance_family_beta(x, mu, phi)
 }
 \description{
 Get distributional variance for beta-family
