+++ man/dot-collapse_cond.Rd
@@ -3,12 +3,14 @@
 \name{.collapse_cond}
 \alias{.collapse_cond}
 \title{glmmTMB returns a list of model information, one for conditional
-and one for zero-inflated part, so here we "unlist" it}
+and one for zero-inflated part, so here we "unlist" it, returning
+only the conditional part.}
 \usage{
 .collapse_cond(x)
 }
 \description{
 glmmTMB returns a list of model information, one for conditional
-and one for zero-inflated part, so here we "unlist" it
+and one for zero-inflated part, so here we "unlist" it, returning
+only the conditional part.
 }
 \keyword{internal}
