+++ man/dot-variance_zinb.Rd
@@ -1,0 +1,14 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/compute_variances.R
+\name{.variance_zinb}
+\alias{.variance_zinb}
+\title{For zero-inflated negative-binomial models, the distributional variance
+is based on Zuur et al. 2012}
+\usage{
+.variance_zinb(model, sig, faminfo, family_var)
+}
+\description{
+For zero-inflated negative-binomial models, the distributional variance
+is based on Zuur et al. 2012
+}
+\keyword{internal}
