+++ man/dot-random_slope_intercept_corr.Rd
@@ -1,0 +1,12 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/compute_variances.R
+\name{.random_slope_intercept_corr}
+\alias{.random_slope_intercept_corr}
+\title{slope-intercept-correlations (rho 01)}
+\usage{
+.random_slope_intercept_corr(vals, x)
+}
+\description{
+slope-intercept-correlations (rho 01)
+}
+\keyword{internal}
