+++ man/dot-random_slope_variance.Rd
@@ -1,0 +1,12 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/compute_variances.R
+\name{.random_slope_variance}
+\alias{.random_slope_variance}
+\title{random slope-variances (tau 11)}
+\usage{
+.random_slope_variance(vals, x)
+}
+\description{
+random slope-variances (tau 11)
+}
+\keyword{internal}
