+++ man/dot-variance_family_nbinom.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/compute_variances.R
-\name{.get_variance_nbinom_family}
-\alias{.get_variance_nbinom_family}
+\name{.variance_family_nbinom}
+\alias{.variance_family_nbinom}
 \title{Get distributional variance for nbinom-family}
 \usage{
-.get_variance_nbinom_family(x, mu, sig, faminfo)
+.variance_family_nbinom(x, mu, sig, faminfo)
 }
 \description{
 Get distributional variance for nbinom-family
