+++ man/dot-variance_family_default.Rd
@@ -1,11 +1,11 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/compute_variances.R
-\name{.get_variance_default}
-\alias{.get_variance_default}
+\name{.variance_family_default}
+\alias{.variance_family_default}
 \title{Get distribution-specific variance for general and
 undefined families / link-functions}
 \usage{
-.get_variance_default(x, mu, verbose)
+.variance_family_default(x, mu, verbose)
 }
 \description{
 Get distribution-specific variance for general and
