+++ man/dot-variance_family_poisson.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/compute_variances.R
-\name{.get_variance_poisson_family}
-\alias{.get_variance_poisson_family}
+\name{.variance_family_poisson}
+\alias{.variance_family_poisson}
 \title{Get distributional variance for poisson-family}
 \usage{
-.get_variance_poisson_family(x, mu, faminfo)
+.variance_family_poisson(x, mu, faminfo)
 }
 \description{
 Get distributional variance for poisson-family
