### 2019-05-07 #1412 @strengejacke
We're planning to submit a package-update to CRAN, which would change the version number. The changes won't affect the paper or review-process, but should we update the version-number here somehow?
