### 2019-05-07 #1412 @alexpghayes
## Review

This review is of the Github `master` branch on 2019-05-07.

`insight` is a well-written and well-documented package. Reading both the source, pkgdown website and tests was a pleasure. The goal of `insight` is to reduce the complexity of working with model objects by providing an easy and intuitive interface to abstracting information from model objects.

My comments are largely about this interface. I think about this primarily in terms of complexity. In some sense `insight` is competing with the original modeling packages -- if the interface to `insight` is: (1) simpler than fiddling with the model objects, and (2) consistent across model objects, then using `insight` reduces the overall complexity of interacting with model objects.

## Interface complexity

My primary concern is that `insight` has a complicated interface. The homepage outlines the interface, which consists of:

- find_algorithm()
- find_formula()
- find_variables()
- find_terms()

- get_data()
- get_priors()
- get_variance()

- find_parameters() / get_parameters()
- find_predictors() / get_predictors()
- find_random() / get_random()
- find_response() / get_response()

With such a broad interface, I had a hard time figuring out the entry point. Some questions I asked myself:

- Where should I start? After some digging it seems like `model_info()` is the key function that returns all the information `insight` provides at once.

- What is the difference between `get_*()` and `find_*()`?

I strongly recommend shrinking this interface. In particular

- find_parameters() / get_parameters()
- find_predictors() / get_predictors()
- find_random() / get_random()
- find_response() / get_response()

providing nearly identical functionality. Also, I struggled to understand the difference between:

- variables and predictors and response
- terms and formula

This meant I had to read the documentation for each of these methods before figuring out which one did what I wanted. These distinctions should be explained clearly on the front page of the pagedown website. The diagram in the JOSS paper helps somewhat but still needs an accompanying narrative.

For greater discussion of interface design and hiding complexity, I highly recommend *A philosophy of software design* by Ousterhoust, which I've been enjoying recently.

## Consistency

While the documentation is very clear about function returns, I had difficulty understanding why behavior differs from different model objects.

- In `find_parameters()`: `aov` objects fit linear models with `conditional` and `random` terms, but instead this information is stored in `between` and `within` fields of the return

- In `find_algorithm()`: the object return differs from different object types, so I still would need to check which elements are in this list before trying to access them. Also, I belive the term `algorithm` should be replaced by `estimator`, which is more appropriate. Also, for Bayesian models, describing the algorithm as "sampling" is too vague.

- In `find_formula()`: how is `random` different from `correlation`? How is `slopes` different from `conditional`? I would have grouped these together. Being somewhat familiar with the model objects under consideration, my impression is that the returns here reflect choices that made the implementation easy for developers, but that do not necessarily provide a generalized interface for users.

On the whole, I get the impression that model components that I conceive of as conceptually similar often end up in different parts of a return object. If I need to understand how `get_parameters()` behaves for both `lm` and `brm` models, it's unclear how beneficial `get_parameters()` is.

## Miscellenea

**Examples**: The documentation shows how a user can extract information from a model, but it doesn't show how the user can use this information to solve larger problems. Some applications showing how the various functions operate together (and how they are distinct) would be greatly beneficial.

- `/R` is organized by method but `/test` is organized by model
- The homepage and README should have links to information about contributing

## Overall impression

My initial impression is that using `insight` is about as complicated as understanding a model object itself. I think the package needs narrative explaining at a conceptual level how it generalizes across models, and return objects with fewer special cases. Both of these are very difficult tasks due to the heterogeneity of models. `insight` may benefit from narrowing in on particular families of models. For instance, it seems like most existing functionality is targeted at GLMMs.
