### 2019-05-07 #1412 @strengejacke
@alexpghayes Thanks for the detailed and comprehensive review! I think your comments are very helpful to make the package more accessible to users. We'll try to address your issues point-by-point and respond here once we think we have a revised version that is sufficient for re-reviewing again.
