### 2019-05-08 61c7fc6
get_variance no longer stops for no re-variance
