+++ R/compute_variances.R
@@ -65,7 +65,7 @@
   obs.terms <- names(nr[nr == stats::nobs(x)])
 
   # Variance of random effects
-  if (component %in% c("random", "all") && isTRUE(no_random_variance)) {
+  if (component %in% c("random", "all") && !isTRUE(no_random_variance)) {
     var.random <- .compute_variance_random(not.obs.terms, x = x, vals = vals)
   }
 
