+++ DESCRIPTION
@@ -22,56 +22,59 @@ Depends:
     stats,
     utils
 Suggests:
-  AER,
-  afex,
-  BayesFactor,
-  betareg,
-  biglm,
-  blme,
-  brms,
-  censReg,
-  crch,
-  coxme,
-  covr,
-  estimatr,
-  feisr,
-  gam,
-  gamm4,
-  gamlss,
-  gbm,
-  gee,
-  geepack,
-  GLMMadaptive,
-  glmmTMB,
-  gmnl,
-  httr,
-  lfe,
-  logistf,
-  MASS,
-  MCMCglmm,
-  mlogit,
-  multgee,
-  lme4,
-  mgcv,
-  nnet,
-  nlme,
-  ordinal,
-  plm,
-  pscl,
-  quantreg,
-  rms,
-  robust,
-  robustbase,
-  robustlmm,
-  rstanarm,
-  rstudioapi,
-  speedglm,
-  splines,
-  survey,
-  survival,
-  truncreg,
-  testthat,
-  VGAM
+    AER,
+    afex,
+    BayesFactor,
+    betareg,
+    biglm,
+    blme,
+    brms,
+    censReg,
+    crch,
+    coxme,
+    covr,
+    estimatr,
+    feisr,
+    gam,
+    gamm4,
+    gamlss,
+    gbm,
+    gee,
+    geepack,
+    GLMMadaptive,
+    glmmTMB,
+    gmnl,
+    httr,
+    lfe,
+    logistf,
+    MASS,
+    MCMCglmm,
+    mlogit,
+    multgee,
+    lme4,
+    mgcv,
+    nnet,
+    nlme,
+    ordinal,
+    plm,
+    pscl,
+    quantreg,
+    rms,
+    robust,
+    robustbase,
+    robustlmm,
+    rstanarm,
+    rstudioapi,
+    speedglm,
+    splines,
+    survey,
+    survival,
+    truncreg,
+    testthat,
+    VGAM,
+    knitr,
+    rmarkdown
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 6.1.1
+VignetteBuilder: knitr
