### 2019-05-11 19dcdf9
revise vignette
