+++ R/find_variables.R
@@ -68,11 +68,19 @@ get_variables_list <- function(f) {
     f_parts
   })
 
-  # remove "1" nad "0" from variables in random effects
+
+  # remove "1" and "0" from variables in random effects
+
   if (obj_has_name(f, "random")) {
     pos <- which(f$random %in% c("1", "0"))
     if (length(pos)) f$random <- f$random[-pos]
   }
+
+  if (obj_has_name(f, "zero_inflated_random")) {
+    pos <- which(f$zero_inflated_random %in% c("1", "0"))
+    if (length(pos)) f$zero_inflated_random <- f$zero_inflated_random[-pos]
+  }
+
 
   # reorder, so response is first
   compact_list(f[c(length(f), 1:(length(f) - 1))])
