### 2019-05-11 ff382cb
ignore tests on CRAN
