+++ man/find_parameters.Rd
@@ -22,7 +22,7 @@ find_parameters(x, flatten = FALSE, ...)
 \item{x}{A fitted model.}
 
 \item{flatten}{Logical, if \code{TRUE}, the values are returned
-as character vector, not as list.}
+as character vector, not as list. Duplicated values are removed.}
 
 \item{...}{Currently not used.}
 
