+++ tests/testthat/test-bigglm.R
@@ -1,6 +1,6 @@
 .runThisTest <- Sys.getenv("RunAllinsightTests") == "yes"
 
-if (.runThisTest && Sys.getenv("USER") != "travis") {
+if (.runThisTest || Sys.getenv("USER") == "travis") {
 
   if (require("testthat") && require("insight") && require("glmmTMB") && require("biglm")) {
     context("insight, model_info")
