### 2019-05-12 f8908db
fix brms-nonlinear issues
