+++ NEWS.md
@@ -3,6 +3,7 @@
 ## Bug fixes
 
 * `find_parameters()` returned priors for `brmsfit`-objects as `$random`-element. Now, `find_parameters()` returns a `$priors`-element (#98).
+* Fixed various issues with non-linear *brms*-models.
 
 # insight 0.3.0
 
