### 2019-05-12 ca12700
update example model
