+++ tests/testthat/test-response_data2.R
@@ -12,16 +12,19 @@ if (suppressWarnings(
   m2 <- glmer(cbind(incidence, size - incidence) ~ period + (1 | herd), data = cbpp, family = binomial)
   m3 <- glm(cbind(incidence, trials) ~ period, data = cbpp, family = binomial)
   m4 <- glm(cbind(incidence, size - incidence) ~ period, data = cbpp, family = binomial)
+  m5 <- glmer(cbind(incidence, size - incidence) ~ (1 | herd), data = cbpp, family = binomial)
 
   test_that("find_response", {
     expect_equal(find_response(m1, combine = TRUE), "cbind(incidence, trials)")
     expect_equal(find_response(m2, combine = TRUE), "cbind(incidence, size - incidence)")
     expect_equal(find_response(m3, combine = TRUE), "cbind(incidence, trials)")
     expect_equal(find_response(m4, combine = TRUE), "cbind(incidence, size - incidence)")
+    expect_equal(find_response(m5, combine = TRUE), "cbind(incidence, size - incidence)")
     expect_equal(find_response(m1, combine = FALSE), c("incidence", "trials"))
     expect_equal(find_response(m2, combine = FALSE), c("incidence", "size"))
     expect_equal(find_response(m3, combine = FALSE), c("incidence", "trials"))
     expect_equal(find_response(m4, combine = FALSE), c("incidence", "size"))
+    expect_equal(find_response(m5, combine = FALSE), c("incidence", "size"))
   })
 
   test_that("get_response", {
@@ -29,6 +32,7 @@ if (suppressWarnings(
     expect_equal(colnames(get_response(m2)), c("incidence", "size"))
     expect_equal(colnames(get_response(m3)), c("incidence", "trials"))
     expect_equal(colnames(get_response(m4)), c("incidence", "size"))
+    expect_equal(colnames(get_response(m5)), c("incidence", "size"))
   })
 
   test_that("get_data", {
@@ -36,6 +40,7 @@ if (suppressWarnings(
     expect_equal(colnames(get_data(m2)), c("cbind(incidence, size - incidence)", "period", "herd", "incidence", "size"))
     get_data(m3)
     get_data(m4)
+    expect_equal(colnames(get_data(m5)), c("cbind(incidence, size - incidence)", "herd", "incidence", "size"))
   })
 
   set.seed(123)
