### 2019-05-13 05976a3
Minor grammatical updates
