+++ R/link_function.R
@@ -72,6 +72,12 @@ link_function.multinom <- function(x, ...) {
 
 
 #' @export
+link_function.BBreg <- function(x, ...) {
+  stats::make.link(link = "logit")$linkfun
+}
+
+
+#' @export
 link_function.gamlss <- function(x, ...) {
   faminfo <- get(x$family[1], asNamespace("gamlss"))()
   faminfo$mu.linkfun
