+++ R/helper_functions.R
@@ -108,7 +108,7 @@ get_fixed_effects <- function(f) {
 # extract random effects from formula
 get_model_random <- function(f, split_nested = FALSE, model) {
 
-  is_special <- inherits(model, c("MCMCglmm", "gee", "LORgee", "felm", "feis", "BFBayesFactor"))
+  is_special <- inherits(model, c("MCMCglmm", "gee", "LORgee", "felm", "feis", "BFBayesFactor", "BBmm"))
 
   if (!requireNamespace("lme4", quietly = TRUE)) {
     stop("To use this function, please install package 'lme4'.")
