+++ R/link_inverse.R
@@ -341,6 +341,12 @@ link_inverse.BBreg <- function(x, ...) {
 
 
 #' @export
+link_inverse.BBmm <- function(x, ...) {
+  stats::make.link("logit")$linkinv
+}
+
+
+#' @export
 link_inverse.clm <- function(x, ...) {
   stats::make.link(get_ordinal_link(x))$linkinv
 }
