+++ R/find_algorithm.R
@@ -95,6 +95,14 @@ find_algorithm.BBreg <- function(x, ...) {
 
 
 #' @export
+find_algorithm.BBmm <- function(x, ...) {
+  method <- parse(text = deparse(x$call, width.cutoff = 500))[[1]]$method
+  if (is.null(method)) method <- "BB-NR"
+  list(algorithm = "extended likelihood", optimizer = method)
+}
+
+
+#' @export
 find_algorithm.biglm <- function(x, ...) {
   list("algorithm" = "OLS")
 }
