+++ R/get_data.R
@@ -178,7 +178,8 @@ get_data.plm <- function(x, ...) {
   cn <- colnames(mf)
   mf <- as.data.frame(lapply(mf, as.vector))
   colnames(mf) <- cn
-  prepare_get_data(x, mf)
+  model_terms <- find_terms(x, effects = "all", component = "all", flatten = TRUE)
+  prepare_get_data(x, mf[, model_terms, drop = FALSE])
 }
 
 
