+++ R/find_predictors.R
@@ -101,7 +101,7 @@ return_vars <- function(f, x) {
   }
 
   # remove constants
-  l <- lapply(l, .remove_values, c("pi", "1", "0"))
+  l <- lapply(l, .remove_values, c(".", "pi", "1", "0"))
   l <- lapply(l, .remove_values, c(0, 1))
   names(l) <- names(f)[!empty_elements]
 
