+++ tests/testthat/test-BBmm.R
@@ -1,129 +1,134 @@
-if (require("testthat") && require("insight") && require("HRQoL")) {
-  context("insight, BBmm")
+.runThisTest <- Sys.getenv("RunAllinsightTests") == "yes"
 
-  set.seed(18)
-  k <- 100
-  m <- 10
-  x <- rnorm(k, 5, 3)
-  j <- runif(k, 0, 5)
-  fac <- sample(letters[1:4], k, TRUE)
+if (.runThisTest && Sys.getenv("USER") != "travis") {
 
-  beta <- c(-10, 2)
-  p <- 1 / (1 + exp(-(beta[1] + beta[2] * x)))
-  phi <- 1.2
+  if (require("testthat") && require("insight") && require("HRQoL")) {
+    context("insight, BBmm")
 
-  y <- rBB(k, m, p, phi)
-  z <- as.factor(rBI(k,4,0.5,2))
+    set.seed(18)
+    k <- 100
+    m <- 10
+    x <- rnorm(k, 5, 3)
+    j <- runif(k, 0, 5)
+    fac <- sample(letters[1:4], k, TRUE)
 
-  dat <- data.frame(y, x, j, fac, z)
+    beta <- c(-10, 2)
+    p <- 1 / (1 + exp(-(beta[1] + beta[2] * x)))
+    phi <- 1.2
 
-  m1 <- BBmm(
-    fixed.formula = y ~ x,
-    random.formula = ~ z,
-    m = m,
-    data = dat
-  )
+    y <- rBB(k, m, p, phi)
+    z <- as.factor(rBI(k,4,0.5,2))
 
-  test_that("model_info", {
-    expect_true(model_info(m1)$is_binomial)
-    expect_true(model_info(m1)$is_betabinomial)
-    expect_true(model_info(m1)$is_mixed)
-  })
+    dat <- data.frame(y, x, j, fac, z)
 
-  test_that("find_predictors", {
-    expect_identical(find_predictors(m1), list(conditional = "x"))
-    expect_identical(find_predictors(m1, flatten = TRUE), "x")
-    expect_identical(find_predictors(m1, effects = "random"), list(random = "z"))
-    expect_identical(
-      find_predictors(m1, effects = "all"),
-      list(
-        conditional = "x",
-        random = "z"
-      )
+    m1 <- BBmm(
+      fixed.formula = y ~ x,
+      random.formula = ~ z,
+      m = m,
+      data = dat
     )
-  })
 
-  test_that("find_random", {
-    expect_identical(find_random(m1), list(random = "z"))
-  })
+    test_that("model_info", {
+      expect_true(model_info(m1)$is_binomial)
+      expect_true(model_info(m1)$is_betabinomial)
+      expect_true(model_info(m1)$is_mixed)
+    })
 
-  test_that("get_random", {
-    expect_equal(get_random(m1), dat[, "z", drop = FALSE])
-  })
-
-  test_that("find_response", {
-    expect_identical(find_response(m1), "y")
-  })
-
-  test_that("get_response", {
-    expect_equal(get_response(m1), dat$y)
-  })
-
-  test_that("get_predictors", {
-    expect_equal(colnames(get_predictors(m1)), "x")
-  })
-
-  test_that("link_inverse", {
-    expect_equal(link_inverse(m1)(.2), plogis(.2), tolerance = 1e-5)
-  })
-
-  test_that("link_function", {
-    expect_equal(link_function(m1)(.2), qlogis(.2), tolerance = 1e-5)
-  })
-
-  test_that("get_data", {
-    expect_equal(nrow(get_data(m1)), 100)
-    expect_equal(colnames(get_data(m1)),  c("y", "x", "z"))
-  })
-
-  test_that("find_formula", {
[…]
