+++ tests/testthat/test-negbin.R
@@ -1,114 +1,119 @@
-if (require("testthat") && require("insight") && require("aod")) {
-  context("insight, negbin")
+.runThisTest <- Sys.getenv("RunAllinsightTests") == "yes"
 
-  data(dja)
-  m1 <- negbin(y ~ group + offset(log(trisk)), ~ village, data = dja)
+if (.runThisTest && Sys.getenv("USER") != "travis") {
 
-  test_that("model_info", {
-    expect_true(model_info(m1)$is_negbin)
-    expect_true(model_info(m1)$is_mixed)
-  })
+  if (require("testthat") && require("insight") && require("aod")) {
+    context("insight, negbin")
 
-  test_that("find_predictors", {
-    expect_identical(find_predictors(m1), list(conditional = c("group", "trisk")))
-    expect_identical(find_predictors(m1, flatten = TRUE), c("group", "trisk"))
-    expect_identical(find_predictors(m1, effects = "random"), list(random = "village"))
-    expect_identical(
-      find_predictors(m1, effects = "all"),
-      list(
-        conditional = c("group", "trisk"),
-        random = "village"
+    data(dja)
+    m1 <- negbin(y ~ group + offset(log(trisk)), ~ village, data = dja)
+
+    test_that("model_info", {
+      expect_true(model_info(m1)$is_negbin)
+      expect_true(model_info(m1)$is_mixed)
+    })
+
+    test_that("find_predictors", {
+      expect_identical(find_predictors(m1), list(conditional = c("group", "trisk")))
+      expect_identical(find_predictors(m1, flatten = TRUE), c("group", "trisk"))
+      expect_identical(find_predictors(m1, effects = "random"), list(random = "village"))
+      expect_identical(
+        find_predictors(m1, effects = "all"),
+        list(
+          conditional = c("group", "trisk"),
+          random = "village"
+        )
       )
-    )
-  })
+    })
 
-  test_that("find_random", {
-    expect_identical(find_random(m1), list(random = "village"))
-  })
+    test_that("find_random", {
+      expect_identical(find_random(m1), list(random = "village"))
+    })
 
-  test_that("get_random", {
-    expect_equal(get_random(m1), dja[, "village", drop = FALSE])
-  })
+    test_that("get_random", {
+      expect_equal(get_random(m1), dja[, "village", drop = FALSE])
+    })
 
-  test_that("find_response", {
-    expect_identical(find_response(m1), "y")
-    expect_identical(find_response(m1, combine = FALSE), "y")
-  })
+    test_that("find_response", {
+      expect_identical(find_response(m1), "y")
+      expect_identical(find_response(m1, combine = FALSE), "y")
+    })
 
-  test_that("get_response", {
-    expect_equal(get_response(m1), dja[, "y"])
-  })
+    test_that("get_response", {
+      expect_equal(get_response(m1), dja[, "y"])
+    })
 
-  test_that("get_predictors", {
-    expect_equal(colnames(get_predictors(m1)), c("group", "trisk"))
-  })
+    test_that("get_predictors", {
+      expect_equal(colnames(get_predictors(m1)), c("group", "trisk"))
+    })
 
-  test_that("link_inverse", {
-    expect_equal(link_inverse(m1)(.2), exp(.2), tolerance = 1e-5)
-  })
+    test_that("link_inverse", {
+      expect_equal(link_inverse(m1)(.2), exp(.2), tolerance = 1e-5)
+    })
 
-  test_that("link_function", {
-    expect_equal(link_function(m1)(.2), log(.2), tolerance = 1e-5)
-  })
+    test_that("link_function", {
[…]
