### 2019-05-18 30e92a5
allow returning sigma for mv-models
