+++ man/find_parameters.Rd
@@ -2,12 +2,28 @@
 % Please edit documentation in R/find_parameters.R
 \name{find_parameters}
 \alias{find_parameters}
+\alias{find_parameters.gam}
+\alias{find_parameters.BBmm}
+\alias{find_parameters.zeroinfl}
+\alias{find_parameters.hurdle}
 \alias{find_parameters.brmsfit}
 \alias{find_parameters.stanreg}
 \alias{find_parameters.stanmvreg}
 \title{Find names of model parameters}
 \usage{
 find_parameters(x, flatten = FALSE, ...)
+
+\method{find_parameters}{gam}(x, component = c("all", "conditional",
+  "smooth_terms"), flatten = FALSE, ...)
+
+\method{find_parameters}{BBmm}(x, effects = c("all", "fixed", "random"),
+  flatten = FALSE, ...)
+
+\method{find_parameters}{zeroinfl}(x, component = c("all", "conditional",
+  "zi", "zero_inflated"), flatten = FALSE, ...)
+
+\method{find_parameters}{hurdle}(x, component = c("all", "conditional",
+  "zi", "zero_inflated"), flatten = FALSE, ...)
 
 \method{find_parameters}{brmsfit}(x, flatten = FALSE,
   parameters = NULL, ...)
@@ -25,6 +41,15 @@ find_parameters(x, flatten = FALSE, ...)
 as character vector, not as list. Duplicated values are removed.}
 
 \item{...}{Currently not used.}
+
+\item{component}{Should all predictor variables, predictor variables for the
+conditional model, the zero-inflated part of the model, the dispersion
+term or the instrumental variables be returned? Applies to models
+with zero-inflated and/or dispersion formula, or to models with instrumental
+variable (so called fixed-effects regressions). May be abbreviated.}
+
+\item{effects}{Should variables for fixed effects, random effects
+or both be returned? Only applies to mixed models. May be abbreviated.}
 
 \item{parameters}{Regular expression pattern that describes the parameters that
 should be returned.}
@@ -50,6 +75,12 @@ Returns the names of model parameters, like they typically
     appear in the \code{summary()} output. For Bayesian models, the parameter
     names equal the column names of the posterior samples after coercion
     from \code{as.data.frame()}.
+}
+\details{
+In most cases when models either return different "effects" (fixed,
+random) or "components" (conditional, zero-inflated, ...), the arguments
+\code{effects} and \code{component} can be used. Not all model classes that
+support these arguments are listed here in the 'Usage' section.
 }
 \examples{
 data(mtcars)
