+++ man/clean_parameters.Rd
@@ -1,0 +1,25 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/clean_parameters.R
+\name{clean_parameters}
+\alias{clean_parameters}
+\title{Get clean names of model paramters}
+\usage{
+clean_parameters(x, ...)
+}
+\arguments{
+\item{x}{A fitted model.}
+}
+\value{
+A data frame with "cleaned" parameter names and information on
+  group or component where parameters belong to.
+}
+\description{
+This function "cleans" names of model parameters by removing
+patterns like \code{"r_"} or \code{"b[]"} (mostly applicable to Stan models)
+and adding columns with information to which group or component parameters
+belong (i.e. fixed or random, count or zero-inflated...)
+}
+\examples{
+model <- download_model("brms_zi_2")
+clean_parameters(model)
+}
