+++ NAMESPACE
@@ -2,6 +2,7 @@
 
 S3method(clean_names,character)
 S3method(clean_names,default)
+S3method(clean_parameters,brmsfit)
 S3method(find_algorithm,BBmm)
 S3method(find_algorithm,BBreg)
 S3method(find_algorithm,Gam)
@@ -370,6 +371,7 @@ S3method(n_obs,zeroinfl)
 S3method(n_obs,zerotrunc)
 export(all_models_equal)
 export(clean_names)
+export(clean_parameters)
 export(download_model)
 export(find_algorithm)
 export(find_formula)
