### 2019-05-19 cb38150
revise docs
