+++ R/find_parameters.R
@@ -75,7 +75,7 @@ find_parameters.default <- function(x, flatten = FALSE, ...) {
 
 #' @export
 find_parameters.data.frame <- function(x, flatten = FALSE, ...) {
-  stop("A data frame is no valid object for this function")
+  stop("A data frame is no valid object for this function.")
 }
 
 
@@ -264,9 +264,9 @@ find_parameters.aovlist <- function(x, flatten = FALSE, ...) {
 
 
 #' @export
-find_parameters.MixMod <- function(x, effects = c("fixed", "random"), component = c("all", "conditional", "zi", "zero_inflated", "dispersion"), flatten = FALSE, ...) {
+find_parameters.MixMod <- function(x, effects = c("all", "fixed", "random"), component = c("all", "conditional", "zi", "zero_inflated"), flatten = FALSE, ...) {
   if (!requireNamespace("lme4", quietly = TRUE)) {
-    stop("To use this function, please install package 'lme4'.")
+    stop("Package 'lme4' required for this function to work. Please install it.")
   }
 
   re.names <- dimnames(lme4::ranef(x))[[2]]
@@ -290,7 +290,6 @@ find_parameters.MixMod <- function(x, effects = c("fixed", "random"), component 
 
   effects <- match.arg(effects)
   component <- match.arg(component)
-
   elements <- .get_elements(effects = effects, component = component)
   l <- compact_list(l[elements])
 
@@ -306,7 +305,7 @@ find_parameters.MixMod <- function(x, effects = c("fixed", "random"), component 
 #' @export
 find_parameters.merMod <- function(x, effects = c("all", "fixed", "random"), flatten = FALSE, ...) {
   if (!requireNamespace("lme4", quietly = TRUE)) {
-    stop("To use this function, please install package 'lme4'.")
+    stop("Package 'lme4' required for this function to work. Please install it.")
   }
 
   l <- compact_list(list(
@@ -329,7 +328,7 @@ find_parameters.merMod <- function(x, effects = c("all", "fixed", "random"), fla
 #' @export
 find_parameters.rlmerMod <- function(x, effects = c("all", "fixed", "random"), flatten = FALSE, ...) {
   if (!requireNamespace("lme4", quietly = TRUE)) {
-    stop("To use this function, please install package 'lme4'.")
+    stop("Package 'lme4' required for this function to work. Please install it.")
   }
 
   l <- compact_list(list(
@@ -352,7 +351,7 @@ find_parameters.rlmerMod <- function(x, effects = c("all", "fixed", "random"), f
 #' @export
 find_parameters.mixed <- function(x, effects = c("all", "fixed", "random"), flatten = FALSE, ...) {
   if (!requireNamespace("lme4", quietly = TRUE)) {
-    stop("To use this function, please install package 'lme4'.")
+    stop("Package 'lme4' required for this function to work. Please install it.")
   }
 
   l <- compact_list(list(
@@ -376,7 +375,7 @@ find_parameters.mixed <- function(x, effects = c("all", "fixed", "random"), flat
 #' @export
 find_parameters.coxme <- function(x, effects = c("all", "fixed", "random"), flatten = FALSE, ...) {
   if (!requireNamespace("lme4", quietly = TRUE)) {
-    stop("To use this function, please install package 'lme4'.")
+    stop("Package 'lme4' required for this function to work. Please install it.")
   }
 
   l <- compact_list(list(
@@ -399,7 +398,7 @@ find_parameters.coxme <- function(x, effects = c("all", "fixed", "random"), flat
[…]
