+++ R/clean_parameters.R
@@ -25,3 +25,46 @@ clean_parameters.brmsfit <- function(x, ...) {
   pars <- find_parameters(x, effects = "all", component = "all", flatten = FALSE)
 
 }
+
+
+#' @export
+clean_parameters.glmmTMB <- function(x, ...) {
+  pars <- find_parameters(x, effects = "all", component = "all", flatten = FALSE)
+
+  l <- lapply(names(pars), function(i) {
+    eff <- if (grepl("random", i, fixed = TRUE))
+      "random"
+    else
+      "fixed"
+
+    com <- if (grepl("zero_inflated", i, fixed = TRUE))
+      "zero_inflated"
+    else
+      "conditional"
+
+    if (eff == "random") {
+      rand_eff <- lapply(names(pars[[i]]), function(j) {
+        data.frame(
+          Parameter = pars[[i]][[j]],
+          Effect = eff,
+          Component = com,
+          Group = j,
+          stringsAsFactors = FALSE,
+          row.names = NULL
+        )
+      })
+      do.call(rbind, rand_eff)
+    } else {
+      data.frame(
+        Parameter = pars[[i]],
+        Effect = eff,
+        Component = com,
+        Group = "",
+        stringsAsFactors = FALSE,
+        row.names = NULL
+      )
+    }
+  })
+
+  do.call(rbind, l)
+}
