### 2019-05-19 4913f09
revise clean_parameters
