+++ R/clean_parameters.R
@@ -9,7 +9,15 @@
 #' @param x A fitted model.
 #'
 #' @return A data frame with "cleaned" parameter names and information on
-#'   group or component where parameters belong to.
+#'   effects, component and group where parameters belong to. To be consistent
+#'   across different models, the returned data frame always has the four
+#'   columns \code{parameter}, \code{effect}, \code{component} and \code{group},
+#'   even for models where some of these columns are not applicable.
+#'
+#' @details The \code{effects} column indicate if a parameter is a \emph{fixed}
+#' or \emph{random} effect. The \code{component} can either be \emph{conditional}
+#' or \emph{zero_inflated}. For models with random effects, the \code{group}
+#' column indicates the grouping factor of the random effects.
 #'
 #' @examples
 #' model <- download_model("brms_zi_2")
@@ -28,7 +36,7 @@ clean_parameters.brmsfit <- function(x, ...) {
 
 
 #' @export
-clean_parameters.glmmTMB <- function(x, ...) {
+clean_parameters.default <- function(x, ...) {
   pars <- find_parameters(x, effects = "all", component = "all", flatten = FALSE)
 
   l <- lapply(names(pars), function(i) {
@@ -45,10 +53,10 @@ clean_parameters.glmmTMB <- function(x, ...) {
     if (eff == "random") {
       rand_eff <- lapply(names(pars[[i]]), function(j) {
         data.frame(
-          Parameter = pars[[i]][[j]],
-          Effect = eff,
-          Component = com,
-          Group = j,
+          parameter = pars[[i]][[j]],
+          effects = eff,
+          component = com,
+          group = j,
           stringsAsFactors = FALSE,
           row.names = NULL
         )
@@ -56,10 +64,10 @@ clean_parameters.glmmTMB <- function(x, ...) {
       do.call(rbind, rand_eff)
     } else {
       data.frame(
-        Parameter = pars[[i]],
-        Effect = eff,
-        Component = com,
-        Group = "",
+        parameter = pars[[i]],
+        effects = eff,
+        component = com,
+        group = "",
         stringsAsFactors = FALSE,
         row.names = NULL
       )
