+++ R/clean_parameters.R
@@ -10,14 +10,18 @@
 #'
 #' @return A data frame with "cleaned" parameter names and information on
 #'   effects, component and group where parameters belong to. To be consistent
-#'   across different models, the returned data frame always has the four
-#'   columns \code{parameter}, \code{effect}, \code{component} and \code{group},
-#'   even for models where some of these columns are not applicable.
+#'   across different models, the returned data frame always has at least three
+#'   columns \code{parameter}, \code{effect} and \code{component}. See 'Details'.
 #'
 #' @details The \code{effects} column indicate if a parameter is a \emph{fixed}
 #' or \emph{random} effect. The \code{component} can either be \emph{conditional}
 #' or \emph{zero_inflated}. For models with random effects, the \code{group}
-#' column indicates the grouping factor of the random effects.
+#' column indicates the grouping factor of the random effects. For multivariate
+#' response models from \pkg{brms} or \pkg{rstanarm}, an additional \emph{response}
+#' column is included, to indicate which parameters belong to which response
+#' formula. Furthermore, for models from \pkg{brms} or \pkg{rstanarm} a
+#' \emph{cleaned_parameter} column is returned that contains "human readable"
+#' parameter names.
 #'
 #' @examples
 #' model <- download_model("brms_zi_2")
@@ -27,12 +31,6 @@ clean_parameters <- function(x, ...) {
   UseMethod("clean_parameters")
 }
 
-
-#' @export
-clean_parameters.brmsfit <- function(x, ...) {
-  pars <- find_parameters(x, effects = "all", component = "all", flatten = FALSE)
-
-}
 
 
 #' @export
@@ -74,5 +72,140 @@ clean_parameters.default <- function(x, ...) {
     }
   })
 
-  do.call(rbind, l)
+  out <- do.call(rbind, l)
+  .remove_empty_columns_from_pars(out)
+}
+
+
+
+#' @export
+clean_parameters.brmsfit <- function(x, ...) {
+  pars <- find_parameters(x, effects = "all", component = "all", flatten = FALSE)
+  is_mv <- is_multivariate(pars)
+
+  if (is_mv) {
+    l <- do.call(
+      rbind,
+      lapply(names(pars), function(i) .get_brms_params(pars[[i]], response = i))
+    )
+  } else {
+    l <- .get_brms_params(pars)
+  }
+
+  out <- do.call(rbind, l)
+  .remove_empty_columns_from_pars(.clean_brms_params(out, is_mv))
+}
+
+
+
+.get_brms_params <- function(pars, response = NA) {
+  lapply(names(pars), function(i) {
+    eff <- if (grepl("random", i, fixed = TRUE))
+      "random"
+    else
+      "fixed"
+
+    com <- if (grepl("zero_inflated", i, fixed = TRUE))
+      "zero_inflated"
+    else if (grepl("sigma", i, fixed = TRUE))
+      "sigma"
+    else if (grepl("priors", i, fixed = TRUE))
+      "priors"
+    else
+      "conditional"
+
+    data.frame(
+      parameter = pars[[i]],
+      effects = eff,
+      component = com,
+      group = "",
+      response = response,
+      stringsAsFactors = FALSE,
+      row.names = NULL
+    )
+  })
+}
+
+
+.clean_brms_params <- function(out, is_mv) {
+
+  out$cleaned_parameter <- out$parameter
+
[…]
