+++ man/clean_parameters.Rd
@@ -20,6 +20,12 @@ This function "cleans" names of model parameters by removing
 patterns like \code{"r_"} or \code{"b[]"} (mostly applicable to Stan models)
 and adding columns with information to which group or component parameters
 belong (i.e. fixed or random, count or zero-inflated...)
+\cr \cr
+The main purpose of this function is to easily filter and select model parameters,
+especially of posterior samples from Stan models, depending on certain
+characteristics. This might be useful when only selective results should
+be reported or results from all parameters should be filtered to return only
+certain results.
 }
 \details{
 The \code{effects} column indicate if a parameter is a \emph{fixed}
