+++ NAMESPACE
@@ -4,6 +4,8 @@ S3method(clean_names,character)
 S3method(clean_names,default)
 S3method(clean_parameters,brmsfit)
 S3method(clean_parameters,default)
+S3method(clean_parameters,stanmvreg)
+S3method(clean_parameters,stanreg)
 S3method(find_algorithm,BBmm)
 S3method(find_algorithm,BBreg)
 S3method(find_algorithm,Gam)
