### 2019-05-20 f028e81
increase test coverage
