+++ tests/testthat/test-mvrstanarm.R
@@ -160,5 +160,145 @@ if (.runThisTest && Sys.getenv("USER") != "travis") {
     test_that("is_multivariate", {
       expect_true(is_multivariate(m1))
     })
+
+    test_that("clean_parameters", {
+      expect_identical(
+        clean_parameters(m1),
+        structure(list(parameter = c("(Intercept)", "year", "sigma",
+        "(Intercept)", "sexf", "year", "sigma", "b[(Intercept) id:1]",
+        "b[(Intercept) id:2]", "b[(Intercept) id:3]", "b[(Intercept) id:4]",
+        "b[(Intercept) id:5]", "b[(Intercept) id:6]", "b[(Intercept) id:7]",
+        "b[(Intercept) id:8]", "b[(Intercept) id:9]", "b[(Intercept) id:10]",
+        "b[(Intercept) id:11]", "b[(Intercept) id:12]", "b[(Intercept) id:13]",
+        "b[(Intercept) id:14]", "b[(Intercept) id:15]", "b[(Intercept) id:16]",
+        "b[(Intercept) id:17]", "b[(Intercept) id:18]", "b[(Intercept) id:19]",
+        "b[(Intercept) id:20]", "b[(Intercept) id:21]", "b[(Intercept) id:22]",
+        "b[(Intercept) id:23]", "b[(Intercept) id:24]", "b[(Intercept) id:25]",
+        "b[(Intercept) id:26]", "b[(Intercept) id:27]", "b[(Intercept) id:28]",
+        "b[(Intercept) id:29]", "b[(Intercept) id:30]", "b[(Intercept) id:31]",
+        "b[(Intercept) id:32]", "b[(Intercept) id:33]", "b[(Intercept) id:34]",
+        "b[(Intercept) id:35]", "b[(Intercept) id:36]", "b[(Intercept) id:37]",
+        "b[(Intercept) id:38]", "b[(Intercept) id:39]", "b[(Intercept) id:40]",
+        "b[(Intercept) id:1]", "b[year id:1]", "b[(Intercept) id:2]",
+        "b[year id:2]", "b[(Intercept) id:3]", "b[year id:3]", "b[(Intercept) id:4]",
+        "b[year id:4]", "b[(Intercept) id:5]", "b[year id:5]", "b[(Intercept) id:6]",
+        "b[year id:6]", "b[(Intercept) id:7]", "b[year id:7]", "b[(Intercept) id:8]",
+        "b[year id:8]", "b[(Intercept) id:9]", "b[year id:9]", "b[(Intercept) id:10]",
+        "b[year id:10]", "b[(Intercept) id:11]", "b[year id:11]", "b[(Intercept) id:12]",
+        "b[year id:12]", "b[(Intercept) id:13]", "b[year id:13]", "b[(Intercept) id:14]",
+        "b[year id:14]", "b[(Intercept) id:15]", "b[year id:15]", "b[(Intercept) id:16]",
+        "b[year id:16]", "b[(Intercept) id:17]", "b[year id:17]", "b[(Intercept) id:18]",
+        "b[year id:18]", "b[(Intercept) id:19]", "b[year id:19]", "b[(Intercept) id:20]",
+        "b[year id:20]", "b[(Intercept) id:21]", "b[year id:21]", "b[(Intercept) id:22]",
+        "b[year id:22]", "b[(Intercept) id:23]", "b[year id:23]", "b[(Intercept) id:24]",
+        "b[year id:24]", "b[(Intercept) id:25]", "b[year id:25]", "b[(Intercept) id:26]",
+        "b[year id:26]", "b[(Intercept) id:27]", "b[year id:27]", "b[(Intercept) id:28]",
[…]
