### 2019-05-21 eba432b
indicate smooth sd
