+++ NEWS.md
@@ -17,6 +17,7 @@
 ## Bug fixes
 
 * `find_parameters()` returned priors for `brmsfit`-objects as `$random`-element. Now, `find_parameters()` returns a `$priors`-element (#98).
+* `find_parameters()` did not remove smooth-parameters that used `te()` or `ti()`.
 * Fixed various issues with non-linear *brms*-models.
 * `find_formula()` (and hence, `find_response()` or `get_data()`) did not work for multi-column responses in null-models (#100).
 * Fix bugs with models from package *plm* that occured during the latest plm-update.
