### 2019-05-22 02d56a2
`model_info()` now returns an element `$is_hurdle` for hurdle models.
