+++ man/model_info.Rd
@@ -39,6 +39,7 @@ Retrieve information from model objects.
      \item \code{is_censored}: model is a censored model
      \item \code{is_zeroinf}: model has zero-inflation component
      \item \code{is_zero_inflated}: alias for \code{is_zeroinf}
+     \item \code{is_hurdle}: model has zero-inflation component and is a hurdle-model (truncated family distribution)
      \item \code{is_mixed}: model is a mixed effects model (with random effects)
      \item \code{is_multivariate}: model is a multivariate response model (currently only works for \emph{brmsfit} objects)
      \item \code{is_trial}: model response contains additional information about the trials
