### 2019-05-22 17b3b7b
support panelr::wbm()
