+++ R/find_predictors.R
@@ -104,6 +104,7 @@ return_vars <- function(f, x) {
   # remove constants
   l <- lapply(l, .remove_values, c(".", "pi", "1", "0"))
   l <- lapply(l, .remove_values, c(0, 1))
+  l <- lapply(l, function(i) gsub("`", "", i, fixed = TRUE))
   names(l) <- names(f)[!empty_elements]
 
   l
