+++ NAMESPACE
@@ -66,6 +66,7 @@ S3method(find_formula,rlmerMod)
 S3method(find_formula,stanmvreg)
 S3method(find_formula,stanreg)
 S3method(find_formula,tobit)
+S3method(find_formula,wbm)
 S3method(find_formula,zeroinfl)
 S3method(find_formula,zerotrunc)
 S3method(find_parameters,BBmm)
@@ -95,6 +96,7 @@ S3method(find_parameters,rqss)
 S3method(find_parameters,stanmvreg)
 S3method(find_parameters,stanreg)
 S3method(find_parameters,vgam)
+S3method(find_parameters,wbm)
 S3method(find_parameters,zeroinfl)
 S3method(find_parameters,zerotrunc)
 S3method(get_data,BBmm)
@@ -132,6 +134,7 @@ S3method(get_data,stanreg)
 S3method(get_data,tobit)
 S3method(get_data,vgam)
 S3method(get_data,vglm)
+S3method(get_data,wbm)
 S3method(get_data,zeroinfl)
 S3method(get_data,zerotrunc)
 S3method(get_parameters,BBmm)
