+++ tests/testthat/test-panelr.R
@@ -4,23 +4,30 @@ if (require("testthat") && require("insight") && require("panelr")) {
   data("WageData")
   wages <- panel_data(WageData, id = id, wave = t)
   m1 <- wbm(lwage ~ lag(union) + wks | blk + fem | blk * lag(union), data = wages)
+  m2 <- wbm(lwage ~ lag(union) + wks | blk + fem | blk * (t | union), data = wages)
 
   test_that("model_info", {
     expect_true(model_info(m1)$is_linear)
+    expect_true(model_info(m2)$is_linear)
   })
 
   test_that("find_predictors", {
     expect_identical(find_predictors(m1), list(conditional = c("union", "wks"), instruments = c("blk", "fem")))
     expect_identical(find_predictors(m1, flatten = TRUE), c("union", "wks", "blk", "fem"))
     expect_identical(find_predictors(m1, effects = "random"), list(random = "lag(union)"))
+
+    expect_identical(find_predictors(m2), list(conditional = c("union", "wks"), instruments = c("blk", "fem")))
+    expect_identical(find_predictors(m2, effects = "random"), list(random = "union"))
   })
 
   test_that("find_random", {
     expect_identical(find_random(m1), list(random = "lag(union)"))
+    expect_identical(find_random(m2), list(random = "union"))
   })
 
   test_that("get_random", {
     expect_equal(get_random(m1)[[1]], wages$union)
+    expect_equal(get_random(m2)[[1]], wages$union)
   })
 
   test_that("find_response", {
@@ -54,11 +61,23 @@ if (require("testthat") && require("insight") && require("panelr")) {
         random = as.formula("~blk | lag(union)")
       )
     )
+
+    expect_equal(
+      find_formula(m2),
+      list(
+        conditional = as.formula("lwage ~ lag(union) + wks"),
+        instruments = as.formula("~blk + fem"),
+        random = as.formula("~blk * t | union")
+      )
+    )
   })
 
   test_that("find_terms", {
     expect_equal(find_terms(m1), list(response = "lwage", conditional = c("union", "wks"), instruments = c("blk", "fem"), random = "lag(union)"))
     expect_equal(find_terms(m1, flatten = TRUE), c("lwage", "union", "wks", "blk", "fem", "lag(union)"))
+
+    expect_equal(find_terms(m2), list(response = "lwage", conditional = c("union", "wks"), instruments = c("blk", "fem"), random = "union"))
+    expect_equal(find_terms(m2, flatten = TRUE), c("lwage", "union", "wks", "blk", "fem"))
   })
 
   test_that("n_obs", {
@@ -73,12 +92,20 @@ if (require("testthat") && require("insight") && require("panelr")) {
     expect_equal(
       find_parameters(m1),
       list(
-        conditional = c("(Intercept)", "imean(lag(union))", "imean(wks)",
-                        "lag(union)", "wks", "blk", "fem", "lag(union):blk"),
-        random = "(Intercept)"
+        conditional = c("(Intercept)", "imean(lag(union))", "imean(wks)", "lag(union)", "wks", "blk", "fem", "lag(union):blk"),
+        random = list(id = "(Intercept)")
       )
     )
     expect_equal(nrow(get_parameters(m1)), 8)
+
+    expect_equal(
+      find_parameters(m2),
+      list(
+        conditional = c("(Intercept)", "imean(lag(union))", "imean(wks)", "lag(union)", "wks", "blk", "fem"),
+        random = list(id = "(Intercept)",
+                      union = c("(Intercept)", "t")))
+    )
+
   })
 
   test_that("find_variables", {
@@ -87,6 +114,11 @@ if (require("testthat") && require("insight") && require("panelr")) {
[…]
