### 2019-05-23 8ef1afe
we need demeaned vars as well in data
