+++ R/get_data.R
@@ -186,13 +186,15 @@ get_data.plm <- function(x, ...) {
 #' @export
 get_data.wbm <- function(x, effects = c("all", "fixed", "random"), ...) {
   effects <- match.arg(effects)
-  dat <- as.data.frame(x@orig_data)
+  mf <- stats::model.frame(x)
+
+  # dat <- as.data.frame(x@orig_data)
 
   mf <- switch(
     effects,
-    all = dat[, unique(clean_names(find_terms(x, flatten = TRUE))), drop = FALSE],
-    fixed = dat[, unique(clean_names(find_terms(x, effects = "fixed", flatten = TRUE))), drop = FALSE],
-    random = dat[, unique(clean_names(find_random(x, flatten = TRUE))), drop = FALSE]
+    all = ,
+    fixed = mf,
+    random = mf[, unique(find_random(x, split_nested = TRUE, flatten = TRUE)), drop = FALSE]
   )
 
   prepare_get_data(x, stats::na.omit(mf))
