+++ R/get_data.R
@@ -197,6 +197,9 @@ get_data.wbm <- function(x, effects = c("all", "fixed", "random"), ...) {
     random = mf[, unique(find_random(x, split_nested = TRUE, flatten = TRUE)), drop = FALSE]
   )
 
+  resp.col <- which(colnames(mf) == find_response(x))
+  mf <- mf[, c(resp.col, 1:ncol(mf)[-resp.col])]
+
   prepare_get_data(x, stats::na.omit(mf))
 }
 
