### 2019-05-23 9f1ff0e
fix panelr issues
