### 2019-05-23 92540ed
update tests and descr
