### 2019-05-23 bfd1319
get_variance() works for wbm()
