+++ NAMESPACE
@@ -373,6 +373,7 @@ S3method(n_obs,multinom)
 S3method(n_obs,rq)
 S3method(n_obs,stanmvreg)
 S3method(n_obs,svyolr)
+S3method(n_obs,wbm)
 S3method(n_obs,zeroinfl)
 S3method(n_obs,zerotrunc)
 export(all_models_equal)
