### 2019-05-27 b68082d
cauchy in lowercase for consistency
