### 2019-05-27 5acab82
parameter name for consistency with get_parameters.BFBayesFactor
