### 2019-05-27 0460c76
fix test issues
