### 2019-05-28 bd23b75
clean_parameters() now finds random smooth effects
