+++ R/compute_variances.R
@@ -1,5 +1,4 @@
 #' @importFrom stats nobs
-#' @keywords internal
 .compute_variances <- function(x, component, name_fun = NULL, name_full = NULL, verbose = TRUE) {
 
   ## Original code taken from GitGub-Repo of package glmmTMB
@@ -119,12 +118,11 @@
 
 
 
-#' store essential information on coefficients, model matrix and so on
-#' as list, since we need these information throughout the functions to
-#' calculate the variance components...
-#'
+# store essential information on coefficients, model matrix and so on
+# as list, since we need these information throughout the functions to
+# calculate the variance components...
+#
 #' @importFrom stats model.matrix
-#' @keywords internal
 .get_variance_information <- function(x, faminfo, name_fun = "get_variances", verbose = TRUE) {
   if (!requireNamespace("lme4", quietly = TRUE)) {
     stop("Package `lme4` needs to be installed to compute variances for mixed models.", call. = FALSE)
@@ -202,9 +200,7 @@
 
 
 
-#' helper-function, telling user if family / distribution is supported or not
-#'
-#' @keywords internal
+# helper-function, telling user if family / distribution is supported or not
 .badlink <- function(link, family, verbose = TRUE) {
   if (verbose) {
     warning(sprintf("Model link '%s' is not yet supported for the %s distribution.", link, family), call. = FALSE)
@@ -215,11 +211,9 @@
 
 
 
-#' glmmTMB returns a list of model information, one for conditional
-#' and one for zero-inflated part, so here we "unlist" it, returning
-#' only the conditional part.
-#'
-#' @keywords internal
+# glmmTMB returns a list of model information, one for conditional
+# and one for zero-inflated part, so here we "unlist" it, returning
+# only the conditional part.
 .collapse_cond <- function(x) {
   if (is.list(x) && "cond" %in% names(x)) {
     x[["cond"]]
@@ -231,10 +225,9 @@
 
 
 
-#' Get fixed effects variance
-#'
+# Get fixed effects variance
+#
 #' @importFrom stats var
-#' @keywords internal
 .compute_variance_fixed <- function(vals) {
   with(vals, stats::var(as.vector(beta %*% t(X))))
 }
@@ -243,10 +236,9 @@
 
 
 
-#' Compute variance associated with a random-effects term (Johnson 2014)
-#'
+# Compute variance associated with a random-effects term (Johnson 2014)
+#
 #' @importFrom stats nobs
-#' @keywords internal
 .compute_variance_random <- function(terms, x, vals) {
 
   sigma_sum <- function(Sigma) {
@@ -275,9 +267,7 @@
 
 
 
-#' Calculate Distribution-specific variance (Nakagawa et al. 2017)
-#'
-#' @keywords internal
+# Calculate Distribution-specific variance (Nakagawa et al. 2017)
 .compute_variance_distribution <- function(x, var.cor, faminfo, name, verbose = TRUE) {
   if (inherits(x, "lme"))
     sig <- x$sigma
@@ -327,9 +317,7 @@
 
 
 
-#' Get dispersion-specific variance
-#'
-#' @keywords internal
+# Get dispersion-specific variance
 .compute_variance_dispersion <- function(x, vals, faminfo, obs.terms) {
   if (faminfo$is_linear) {
     0
@@ -345,12 +333,11 @@
 
 
 
-#' This is the core-function to calculate the distribution-specific variance
-#' Nakagawa et al. 2017 propose three different methods, here we only rely
-#' on the lognormal-approximation.
-#'
+# This is the core-function to calculate the distribution-specific variance
[…]
