### 2019-05-28 aaceeee
don't run tests on CRAN
