+++ tests/testthat/test-GLMMadaptive.R
@@ -1,171 +1,172 @@
-if (require("testthat") && require("insight") && require("GLMMadaptive") && require("lme4")) {
-  context("insight, model_info")
+.runThisTest <- Sys.getenv("RunAllinsightTests") == "yes"
 
-  fish <- read.csv("https://stats.idre.ucla.edu/stat/data/fish.csv")
-  fish$nofish <- as.factor(fish$nofish)
-  fish$livebait <- as.factor(fish$livebait)
-  fish$camper <- as.factor(fish$camper)
+if (.runThisTest || Sys.getenv("USER") == "travis") {
 
-  m <- download_model("GLMMadaptive_zi_2")
-  m2 <- download_model("GLMMadaptive_zi_1")
+  if (require("testthat") && require("insight") && require("GLMMadaptive") && require("lme4")) {
+    context("insight, model_info")
 
-  data(cbpp)
-  m3 <- GLMMadaptive::mixed_model(
-    cbind(incidence, size - incidence) ~ period,
-    random = ~ 1 | herd,
-    data = cbpp,
-    family = binomial
-  )
+    m <- download_model("GLMMadaptive_zi_2")
+    m2 <- download_model("GLMMadaptive_zi_1")
 
-  test_that("model_info", {
-    expect_true(model_info(m)$is_zeroinf)
-    expect_true(model_info(m)$is_count)
-    expect_true(model_info(m)$is_pois)
-    expect_false(model_info(m)$is_negbin)
-  })
+    data(cbpp)
+    m3 <- GLMMadaptive::mixed_model(
+      cbind(incidence, size - incidence) ~ period,
+      random = ~ 1 | herd,
+      data = cbpp,
+      family = binomial
+    )
 
-  test_that("find_predictors", {
-    expect_identical(find_predictors(m, effects = "fixed")$conditional, c("child", "camper"))
-    expect_identical(find_predictors(m, effects = "fixed")$zero_inflated, c("child", "livebait"))
-    expect_identical(find_predictors(m, effects = "all", flatten = TRUE), c("child", "camper", "persons", "livebait"))
-    expect_identical(find_predictors(m, effects = "all")$zero_inflated_random, c("persons"))
-    expect_identical(find_predictors(m, effects = "random")$random, "persons")
-    expect_identical(find_predictors(m, effects = "fixed", component = "cond", flatten = TRUE), c("child", "camper"))
-    expect_identical(find_predictors(m, effects = "all", component = "cond", flatten = TRUE), c("child", "camper", "persons"))
-    expect_identical(find_predictors(m, effects = "all", component = "cond")$conditional, c("child", "camper"))
+    test_that("model_info", {
+      expect_true(model_info(m)$is_zeroinf)
+      expect_true(model_info(m)$is_count)
+      expect_true(model_info(m)$is_pois)
+      expect_false(model_info(m)$is_negbin)
+    })
 
-    expect_identical(find_predictors(m, effects = "random", component = "cond", flatten = TRUE), "persons")
-    expect_identical(find_predictors(m, effects = "fixed", component = "zi", flatten = TRUE), c("child", "livebait"))
-    expect_identical(find_predictors(m, effects = "all", component = "zi", flatten = TRUE), c("child", "livebait", "persons"))
-    expect_identical(find_predictors(m, effects = "random", component = "zi", flatten = TRUE), "persons")
-    expect_null(find_predictors(m, effects = "fixed", component = "dispersion", flatten = TRUE))
-    expect_null(find_predictors(m, effects = "all", component = "dispersion", flatten = TRUE))
-    expect_null(find_predictors(m, effects = "random", component = "dispersion", flatten = TRUE))
-  })
+    test_that("find_predictors", {
+      expect_identical(find_predictors(m, effects = "fixed")$conditional, c("child", "camper"))
[…]
