+++ tests/testthat/test-rstanarm.R
@@ -1,5 +1,6 @@
 .runThisTest <- Sys.getenv("RunAllinsightTests") == "yes"
-# if (.runThisTest) {
+
+if (.runThisTest || Sys.getenv("USER") == "travis") {
 
   if (suppressWarnings(
     require("testthat") &&
@@ -166,4 +167,4 @@
 
   }
 
-# }
+}
