+++ NAMESPACE
@@ -382,6 +382,7 @@ export(clean_parameters)
 export(download_model)
 export(find_algorithm)
 export(find_formula)
+export(find_interactions)
 export(find_parameters)
 export(find_predictors)
 export(find_random)
