### 2019-05-28 7b347e4
include option to get also main effects, maybe for later use
