### 2019-05-29 03116b3
more clear warnings when data is not in env
