### 2019-06-02 24fcbdf
Fixed issues with `all_model_equal()`
