+++ R/find_formula.R
@@ -56,7 +56,7 @@ find_formula.default <- function(x, ...) {
 #' @export
 find_formula.gls <- function(x, ...) {
   ## TODO this is an intermediate fix to return the correlation variables from gls-objects
-  f_corr <- parse(text = deparse(x$call$correlation, width.cutoff = 500))[[1]]$form
+  f_corr <- parse(text = .safe_deparse(x$call$correlation))[[1]]$form
 
   l <- tryCatch({
     list(
@@ -107,11 +107,11 @@ find_formula.gamm <- function(x, ...) {
 #' @export
 find_formula.gee <- function(x, ...) {
   tryCatch({
-    id <- parse(text = deparse(x$call, width.cutoff = 500))[[1]]$id
+    id <- parse(text = .safe_deparse(x$call))[[1]]$id
 
     # alternative regex-patterns that also work:
-    # sub(".*id ?= ?(.*?),.*", "\\1", deparse(x$call, width.cutoff = 500), perl = TRUE)
-    # sub(".*\\bid\\s*=\\s*([^,]+).*", "\\1", deparse(x$call, width.cutoff = 500), perl = TRUE)
+    # sub(".*id ?= ?(.*?),.*", "\\1", .safe_deparse(x$call), perl = TRUE)
+    # sub(".*\\bid\\s*=\\s*([^,]+).*", "\\1", .safe_deparse(x$call), perl = TRUE)
 
     list(
       conditional = stats::formula(x),
@@ -128,11 +128,11 @@ find_formula.gee <- function(x, ...) {
 #' @export
 find_formula.LORgee <- function(x, ...) {
   tryCatch({
-    id <- parse(text = deparse(x$call, width.cutoff = 500))[[1]]$id
+    id <- parse(text = .safe_deparse(x$call))[[1]]$id
 
     # alternative regex-patterns that also work:
-    # sub(".*id ?= ?(.*?),.*", "\\1", deparse(x$call, width.cutoff = 500), perl = TRUE)
-    # sub(".*\\bid\\s*=\\s*([^,]+).*", "\\1", deparse(x$call, width.cutoff = 500), perl = TRUE)
+    # sub(".*id ?= ?(.*?),.*", "\\1", .safe_deparse(x$call), perl = TRUE)
+    # sub(".*\\bid\\s*=\\s*([^,]+).*", "\\1", .safe_deparse(x$call), perl = TRUE)
 
     list(
       conditional = stats::formula(x),
@@ -149,7 +149,7 @@ find_formula.LORgee <- function(x, ...) {
 #' @export
 find_formula.ivreg <- function(x, ...) {
   tryCatch({
-    f <- deparse(stats::formula(x), width.cutoff = 500)
+    f <- .safe_deparse(stats::formula(x))
     cond <- trim(substr(f, start = 0, stop = regexpr(pattern = "\\|", f) - 1))
     instr <- trim(substr(f, regexpr(pattern = "\\|", f) + 1, stop = 10000L))
 
@@ -168,7 +168,7 @@ find_formula.ivreg <- function(x, ...) {
 #' @export
 find_formula.iv_robust <- function(x, ...) {
   tryCatch({
-    f <- deparse(stats::formula(x), width.cutoff = 500)
+    f <- .safe_deparse(stats::formula(x))
     cond <- trim(gsub("(.*)\\+(\\s)*\\((.*)\\)", "\\1", f))
     instr <- trim(gsub("(.*)\\((.*)\\)", "\\2", f))
 
@@ -187,7 +187,7 @@ find_formula.iv_robust <- function(x, ...) {
 #' @export
 find_formula.plm <- function(x, ...) {
   tryCatch({
[…]
