### 2019-06-04 0f6f400
Significantly reworked front end
Addressed many points on the front end, specifically dealing with: 
1. a clearer statement of need
2. clarification of differences between `get_*` and `find_*`
3. new section and front end addition for "entry point"/where to start for the modal user (i.e., with `model_info()`, per the reviewer's comments)
4. edited/cleaned discussion throughout to read more cleanly (hopefully...)
