+++ R/clean_parameters.R
@@ -7,10 +7,10 @@
 #' belong (i.e. fixed or random, count or zero-inflated...)
 #' \cr \cr
 #' The main purpose of this function is to easily filter and select model parameters,
-#' especially of posterior samples from Stan models, depending on certain
-#' characteristics. This might be useful when only selective results should
-#' be reported or results from all parameters should be filtered to return only
-#' certain results.
+#' in particular of - but not limited to - posterior samples from Stan models,
+#' depending on certain characteristics. This might be useful when only selective
+#' results should be reported or results from all parameters should be filtered
+#' to return only certain results.
 #'
 #' @param x A fitted model.
 #' @inheritParams find_parameters
