### 2019-06-04 7f53bfc
update vignette accordingly, minor changes in paper
