+++ man/get_parameters.Rd
@@ -9,6 +9,7 @@
 \alias{get_parameters.Gam}
 \alias{get_parameters.zeroinfl}
 \alias{get_parameters.gamm}
+\alias{get_parameters.aovlist}
 \alias{get_parameters.hurdle}
 \alias{get_parameters.coxme}
 \alias{get_parameters.merMod}
@@ -27,7 +28,8 @@ get_parameters(x, ...)
 
 \method{get_parameters}{BBmm}(x, effects = c("fixed", "random"), ...)
 
-\method{get_parameters}{glimML}(x, effects = c("fixed", "random"), ...)
+\method{get_parameters}{glimML}(x, effects = c("fixed", "random", "all"),
+  ...)
 
 \method{get_parameters}{gam}(x, component = c("all", "conditional",
   "smooth_terms"), ...)
@@ -43,6 +45,9 @@ get_parameters(x, ...)
 
 \method{get_parameters}{gamm}(x, component = c("all", "conditional",
   "smooth_terms"), ...)
+
+\method{get_parameters}{aovlist}(x, effects = c("fixed", "random",
+  "all"), ...)
 
 \method{get_parameters}{hurdle}(x, component = c("all", "conditional",
   "zi", "zero_inflated"), ...)
@@ -105,7 +110,7 @@ should be returned.}
   \item for non-Bayesian models and if \code{effects = "fixed"}, a data frame with two columns: the parameter names and the related point estimates
   \item if \code{effects = "random"}, a list of data frames with the random effects (as returned by \code{ranef()}
   \item for Bayesian models, the posterior samples from the requested parameters as data frame.
-  \item for Anova (\code{aov()}) with error term, a list of parameters for the conditional, the within-subject and the between-subjects parameters
+  \item for Anova (\code{aov()}) with error term, a list of parameters for the conditional and the random effects parameters
   \item for models with smooth terms or zero-inflation component, a data frame with three columns: the parameter names, the related point estimates and the component
 }
 }
