### 2019-06-05 70a3c58
update list of supported models
