### 2019-06-05 2ad8a0c
add pagebreaks, to proper assign figures to paragraphs
