### 2019-06-05 3f36961
add reference to figures
