### 2019-06-05 #1412 @strengejacke
@karthik Any ideas why compilation fails? It seems to be a code-block in markdown (`x + I(x^2)`) which causes the error, although when I build the PDF locally, it works fine.
