+++ R/get_data.R
@@ -39,7 +39,7 @@ get_data.default <- function(x, ...) {
 
   mf <- tryCatch({
     if (inherits(x, "Zelig-relogit")) {
-      get_zelig_relogit_frame(x)
+      .get_zelig_relogit_frame(x)
     } else {
       stats::model.frame(x)
     }
@@ -49,21 +49,21 @@ get_data.default <- function(x, ...) {
   }
   )
 
-  prepare_get_data(x, mf)
+  .prepare_get_data(x, mf)
 }
 
 
 #' @export
 get_data.biglm <- function(x, ...) {
   mf <- stats::model.frame(x)
-  prepare_get_data(x, mf)
+  .prepare_get_data(x, mf)
 }
 
 
 #' @export
 get_data.bigglm <- function(x, ...) {
   mf <- stats::model.frame(x)
-  prepare_get_data(x, mf)
+  .prepare_get_data(x, mf)
 }
 
 
@@ -104,7 +104,7 @@ get_data.LORgee <- function(x, effects = c("all", "fixed", "random"), ...) {
     stats::model.frame(x)
   })
 
-  prepare_get_data(x, stats::na.omit(mf))
+  .prepare_get_data(x, stats::na.omit(mf))
 }
 
 
@@ -124,7 +124,7 @@ get_data.BBmm <- function(x, effects = c("all", "fixed", "random"), ...) {
     x$X
   })
 
-  prepare_get_data(x, stats::na.omit(mf))
+  .prepare_get_data(x, stats::na.omit(mf))
 }
 
 
@@ -139,7 +139,7 @@ get_data.glimML <- function(x, effects = c("all", "fixed", "random"), ...) {
     random = dat[, find_random(x, flatten = TRUE), drop = FALSE]
   )
 
-  prepare_get_data(x, stats::na.omit(mf))
+  .prepare_get_data(x, stats::na.omit(mf))
 }
 
 
@@ -162,7 +162,7 @@ get_data.tobit <- function(x, ...) {
   ft <- find_terms(x, flatten = TRUE)
   remain <- intersect(ft, colnames(dat))
 
-  prepare_get_data(x, stats::na.omit(dat[, remain, drop = FALSE]))
+  .prepare_get_data(x, stats::na.omit(dat[, remain, drop = FALSE]))
 }
 
 
@@ -179,7 +179,7 @@ get_data.plm <- function(x, ...) {
   mf <- as.data.frame(lapply(mf, as.vector))
   colnames(mf) <- clean_names(cn)
   model_terms <- find_terms(x, effects = "all", component = "all", flatten = TRUE)
-  prepare_get_data(x, mf[, model_terms, drop = FALSE])
+  .prepare_get_data(x, mf[, model_terms, drop = FALSE])
 }
 
 
@@ -197,7 +197,7 @@ get_data.wbm <- function(x, effects = c("all", "fixed", "random"), ...) {
   resp.col <- which(colnames(mf) == find_response(x))
   mf <- mf[, c(resp.col, (1:ncol(mf))[-resp.col])]
 
-  prepare_get_data(x, stats::na.omit(mf))
+  .prepare_get_data(x, stats::na.omit(mf))
 }
 
 
@@ -206,7 +206,7 @@ get_data.gamm <- function(x, ...) {
   x <- x$gam
   class(x) <- c(class(x), c("glm", "lm"))
   mf <- stats::model.frame(x)
-  prepare_get_data(x, mf)
+  .prepare_get_data(x, mf)
 }
 
 
@@ -231,7 +231,7 @@ get_data.ivreg <- function(x, ...) {
     )
   }
 
-  prepare_get_data(x, stats::na.omit(final_mf))
[…]
