+++ R/clean_parameters.R
@@ -322,12 +322,12 @@ clean_parameters.stanmvreg <- function(x, ...) {
     x <- x[, -pos]
   }
 
-  if (obj_has_name(x, "group") && is_empty_string(x$group)) {
+  if (obj_has_name(x, "group") && .is_empty_string(x$group)) {
     pos <- which(colnames(x) == "group")
     x <- x[, -pos]
   }
 
-  if (obj_has_name(x, "fun") && is_empty_string(x$fun)) {
+  if (obj_has_name(x, "fun") && .is_empty_string(x$fun)) {
     pos <- which(colnames(x) == "fun")
     x <- x[, -pos]
   }
