### 2019-06-05 bad4a27
update paper refs
