+++ paper/paper.bib
@@ -1,8 +1,50 @@
-  @Manual{rcore,
-    title = {R: A Language and Environment for Statistical Computing},
-    author = {{R Core Team}},
-    organization = {R Foundation for Statistical Computing},
-    address = {Vienna, Austria},
-    year = {2019},
-    url = {https://www.R-project.org/},
-  }
+@Manual{rcore,
+  title = {R: A Language and Environment for Statistical Computing},
+  author = {{R Core Team}},
+  organization = {R Foundation for Statistical Computing},
+  address = {Vienna, Austria},
+  year = {2019},
+  url = {https://www.R-project.org/},
+}
+
+@Manual{pkg_performance,
+  title = {performance: Assessment of Regression Models Performance},
+  author = {Daniel Lüdecke and Dominique Makowski and Philip Waggoner},
+  year = {2019},
+  note = {R package version 0.2.0},
+  url = {https://easystats.github.io/performance/},
+}
+
+@Manual{pkg_sjPlot,
+  title = {sjPlot: Data Visualization for Statistics in Social Science},
+  author = {Daniel Lüdecke},
+  year = {2019},
+  note = {R package version 2.6.3},
+  url = {https://CRAN.R-project.org/package=sjPlot},
+  doi = {10.5281/zenodo.1308157},
+}
+
+@Article{pkg_bayestestR,
+  title = {Understand and Describe Bayesian Models and Posterior Distributions using bayestestR},
+  author = {Dominique Makowski and Mattan S. Ben-Shachar and Daniel Lüdecke},
+  journal = {CRAN},
+  doi = {10.5281/zenodo.2556486},
+  year = {2019},
+  note = {R package},
+  url = {https://github.com/easystats/bayestestR},
+}
+
+@article{ludecke_ggeffects_2018,
+	title = {ggeffects: Tidy Data Frames of Marginal Effects from Regression Models},
+	volume = {3},
+	issn = {2475-9066},
+	url = {http://joss.theoj.org/papers/10.21105/joss.00772},
+	doi = {10.21105/joss.00772},
+	number = {26},
+	urldate = {2019-06-05},
+	journal = {Journal of Open Source Software},
+	author = {Lüdecke, Daniel},
+	month = jun,
+	year = {2018},
+	pages = {772}
+}
