### 2019-06-05 f5db53b
add paragraph on use case examples
