+++ paper/paper.md
@@ -154,6 +154,18 @@ find_formula(sample3)
 #> Sepal.Width ~ Species * Petal.Length
 ```
 
+## Examples of Use Cases in R Packages
+
+**insight** is already used by different packages to solve problems that typically occur when the users' inputs are different model objects of varying complexity.
+
+For example, **ggeffects** [@ludecke_ggeffects_2018], a package that computes and visualizes marginal effects of regression models, needs to extract the data (`get_data()`) that was used to fit the models and to retrieve all model predictors (`find_predictors()`) to decide, which covariates are held constant when computing marginal effects. All these information is required in order to create a data frame for `predict(newdata=<data frame>)`. Furthermore, the models' link-functions (`link_function()`) resp. link-inverse-functions (`link_inverse()`) are required to get predictons at the model's response scale.
+
+The **sjPlot**-package [@pkg_sjPlot] creates plots or summary tables from regression models, and uses **insight**-functions to get model-information (`model_info()` or `find_response()`), which is used to build the components of the final plot or table. These information help, for example, to label table columns by providing information on the effect type (odds ratio, incidence rate ratio, ...) or the different model components (which split plots and tables into the "conditional" and "zero-inflated" part of a model, in case of models with zero-inflation).
+
+**bayestestR** [@pkg_bayestestR] mainly relies on `get_priors()` and `get_parameters()` to retrieve the necessary information to compute various indices or statistics of Bayesian models (like HDI, Credible Interval, MCSE, effective sample size, Bayes factors...). The advantage of `get_parameters()` in this context is, for instance, that, no matter, how many parameters the posterior distribution has because of the model's complexity, the necessary data can be accessed easily from the model objects. There is no need to write own complicated code (and regular expressions).
+
+A last example is the **performance**-package [@pkg_performance], which provides functions for computing measures to assess model quality. Many of these indices (e.g. check for overdispersion or zero-inflation, predictive accuracy, logloss, RMSE, ...) require the number of observations (`n_obs()`) or the data from the response-variable (`get_response()`). Again, in this context, functions from **insight** are helpful, because they offer a unified access to these information.
+
 ## Supported Models
 
[…]
