### 2019-06-05 6832ba4
slightly revise docs, mark internal functions
