+++ R/get_parameters.R
@@ -13,7 +13,7 @@
 #'
 #' @return \itemize{
 #'   \item for non-Bayesian models and if \code{effects = "fixed"}, a data frame with two columns: the parameter names and the related point estimates
-#'   \item if \code{effects = "random"}, a list of data frames with the random effects (as returned by \code{ranef()})
+#'   \item if \code{effects = "random"}, a list of data frames with the random effects (as returned by \code{ranef()}), unless the random effects have the same simplified structure as fixed effects (e.g. for models from \pkg{MCMCglmm})
 #'   \item for Bayesian models, the posterior samples from the requested parameters as data frame
 #'   \item for Anova (\code{aov()}) with error term, a list of parameters for the conditional and the random effects parameters
 #'   \item for models with smooth terms or zero-inflation component, a data frame with three columns: the parameter names, the related point estimates and the component
@@ -21,8 +21,7 @@
 #'
 #' @details In most cases when models either return different "effects" (fixed,
 #' random) or "components" (conditional, zero-inflated, ...), the arguments
-#' \code{effects} and \code{component} can be used. Not all model classes that
-#' support these arguments are listed here in the 'Usage' section.
+#' \code{effects} and \code{component} can be used.
 #' \cr \cr
 #' \code{get_parameters()} is comparable to \code{coef()}, however, the coefficients
 #' are returned as data frame (with columns for names and point estimates of
@@ -38,6 +37,7 @@
 get_parameters <- function(x, ...) {
   UseMethod("get_parameters")
 }
+
 
 
 #' @export
@@ -78,6 +78,7 @@ get_parameters.default <- function(x, ...) {
 }
 
 
+
 #' @export
 get_parameters.gbm <- function(x, ...) {
   s <- summary(x, plotit = FALSE)
@@ -90,6 +91,7 @@ get_parameters.gbm <- function(x, ...) {
 }
 
 
+
 #' @export
 get_parameters.BBreg <- function(x, ...) {
   pars <- summary(x)$coefficients
@@ -100,6 +102,7 @@ get_parameters.BBreg <- function(x, ...) {
     row.names = NULL
   )
 }
+
 
 
 #' @rdname get_parameters
@@ -125,6 +128,7 @@ get_parameters.BBmm <- function(x, effects = c("fixed", "random"), ...) {
     l$random
   }
 }
+
 
 
 #' @rdname get_parameters
@@ -166,6 +170,7 @@ get_parameters.glimML <- function(x, effects = c("fixed", "random", "all"), ...)
 }
 
 
+
 #' @export
 get_parameters.gamlss <- function(x, ...) {
   pars <- list(
@@ -190,6 +195,7 @@ get_parameters.gamlss <- function(x, ...) {
 }
 
 
+
 #' @export
 get_parameters.lrm <- function(x, ...) {
   tryCatch({
@@ -208,6 +214,7 @@ get_parameters.lrm <- function(x, ...) {
 }
 
 
+
 #' @export
 get_parameters.aov <- function(x, ...) {
   cf <- stats::coef(x)
@@ -220,10 +227,12 @@ get_parameters.aov <- function(x, ...) {
 }
 
 
+
 #' @export
 get_parameters.data.frame <- function(x, ...) {
   stop("A data frame is no valid object for this function")
 }
+
 
 
 #' @rdname get_parameters
@@ -236,12 +245,13 @@ get_parameters.gam <- function(x, component = c("all", "conditional", "smooth_te
   smooth_terms <- st[, 1]
   names(smooth_terms) <- row.names(st)
 
[…]
