### 2019-06-05 aa67cb5
mark internal functions
