+++ R/find_formula.R
@@ -37,6 +37,7 @@ find_formula <- function(x, ...) {
 }
 
 
+
 #' @export
 find_formula.default <- function(x, ...) {
   if (inherits(x, "list") && obj_has_name(x, "gam")) {
@@ -53,30 +54,13 @@ find_formula.default <- function(x, ...) {
   )
 }
 
-#' @export
-find_formula.gls <- function(x, ...) {
-  ## TODO this is an intermediate fix to return the correlation variables from gls-objects
-  f_corr <- parse(text = .safe_deparse(x$call$correlation))[[1]]$form
-
-  l <- tryCatch({
-    list(
-      conditional = stats::formula(x),
-      correlation = stats::as.formula(f_corr)
-    )
-  },
-  error = function(x) {
-    NULL
-  }
-  )
-
-  compact_list(l)
-}
 
 
 #' @export
 find_formula.data.frame <- function(x, ...) {
   stop("A data frame is no valid object for this function")
 }
+
 
 
 #' @export
@@ -96,12 +80,14 @@ find_formula.gamlss <- function(x, ...) {
 }
 
 
+
 #' @export
 find_formula.gamm <- function(x, ...) {
   x <- x$gam
   class(x) <- c(class(x), c("glm", "lm"))
   NextMethod()
 }
+
 
 
 #' @export
@@ -125,6 +111,28 @@ find_formula.gee <- function(x, ...) {
 }
 
 
+
+#' @export
+find_formula.gls <- function(x, ...) {
+  ## TODO this is an intermediate fix to return the correlation variables from gls-objects
+  f_corr <- parse(text = .safe_deparse(x$call$correlation))[[1]]$form
+
+  l <- tryCatch({
+    list(
+      conditional = stats::formula(x),
+      correlation = stats::as.formula(f_corr)
+    )
+  },
+  error = function(x) {
+    NULL
+  }
+  )
+
+  compact_list(l)
+}
+
+
+
 #' @export
 find_formula.LORgee <- function(x, ...) {
   tryCatch({
@@ -146,6 +154,7 @@ find_formula.LORgee <- function(x, ...) {
 }
 
 
+
 #' @export
 find_formula.ivreg <- function(x, ...) {
   tryCatch({
@@ -165,6 +174,7 @@ find_formula.ivreg <- function(x, ...) {
 }
 
 
+
 #' @export
 find_formula.iv_robust <- function(x, ...) {
   tryCatch({
@@ -182,6 +192,7 @@ find_formula.iv_robust <- function(x, ...) {
   }
   )
 }
+
 
 
 #' @export
@@ -216,6 +227,7 @@ find_formula.plm <- function(x, ...) {
 }
 
 
+
 #' @export
 find_formula.coxme <- function(x, ...) {
   if (!requireNamespace("lme4", quietly = TRUE)) {
@@ -240,6 +252,7 @@ find_formula.coxme <- function(x, ...) {
     random = f.random
   ))
 }
+
 
 
 #' @export
@@ -276,6 +289,7 @@ find_formula.felm <- function(x, ...) {
 }
 
 
+
 #' @export
 find_formula.feis <- function(x, ...) {
   f <- .safe_deparse(stats::formula(x))
@@ -300,6 +314,7 @@ find_formula.feis <- function(x, ...) {
     random = stats::as.formula(paste0("~", id))
   ))
 }
+
 
 
 #' @export
@@ -338,6 +353,7 @@ find_formula.wbm <- function(x, ...) {
 }
 
 
+
 #' @export
 find_formula.BBmm <- function(x, ...) {
[…]
