### 2019-06-05 0e42694
mark internal functions
