+++ R/helper_functions.R
@@ -2,7 +2,7 @@
 trim <- function(x) gsub("^\\s+|\\s+$", "", x)
 
 # remove NULL elements from lists
-compact_list <- function(x) x[!sapply(x, function(i) length(i) == 0 || is.null(i) || any(i == "NULL"))]
+.compact_list <- function(x) x[!sapply(x, function(i) length(i) == 0 || is.null(i) || any(i == "NULL"))]
 
 
 # remove values from vector
