+++ R/find_formula.R
@@ -128,7 +128,7 @@ find_formula.gls <- function(x, ...) {
   }
   )
 
-  compact_list(l)
+  .compact_list(l)
 }
 
 
@@ -247,7 +247,7 @@ find_formula.coxme <- function(x, ...) {
 
   f.cond <- stats::as.formula(get_fixed_effects(f.cond))
 
-  compact_list(list(
+  .compact_list(list(
     conditional = f.cond,
     random = f.random
   ))
@@ -280,7 +280,7 @@ find_formula.felm <- function(x, ...) {
     f.clus <- NULL
   }
 
-  compact_list(list(
+  .compact_list(list(
     conditional = stats::as.formula(f.cond),
     random = stats::as.formula(f.rand),
     instruments = stats::as.formula(f.instr),
@@ -308,7 +308,7 @@ find_formula.feis <- function(x, ...) {
     f.slopes <- NULL
   }
 
-  compact_list(list(
+  .compact_list(list(
     conditional = stats::as.formula(f.cond),
     slopes = stats::as.formula(f.slopes),
     random = stats::as.formula(paste0("~", id))
@@ -345,7 +345,7 @@ find_formula.wbm <- function(x, ...) {
     f.rand <- NULL
   }
 
-  compact_list(list(
+  .compact_list(list(
     conditional = stats::as.formula(f.cond),
     instruments = stats::as.formula(f.instr),
     random = stats::as.formula(f.rand)
@@ -359,7 +359,7 @@ find_formula.BBmm <- function(x, ...) {
   f.cond <- parse(text = .safe_deparse(x$call))[[1]]$fixed.formula
   f.rand <- parse(text = .safe_deparse(x$call))[[1]]$random.formula
 
-  compact_list(list(
+  .compact_list(list(
     conditional = stats::as.formula(f.cond),
     random = stats::as.formula(f.rand)
   ))
@@ -369,7 +369,7 @@ find_formula.BBmm <- function(x, ...) {
 
 #' @export
 find_formula.glimML <- function(x, ...) {
-  compact_list(list(
+  .compact_list(list(
     conditional = x@formula,
     random = x@random
   ))
@@ -473,7 +473,7 @@ find_formula.glmmTMB <- function(x, ...) {
   f.cond <- stats::as.formula(get_fixed_effects(f.cond))
   if (!is.null(f.zi)) f.zi <- stats::as.formula(get_fixed_effects(f.zi))
 
-  compact_list(list(
+  .compact_list(list(
     conditional = f.cond,
     random = f.random,
     zero_inflated = f.zi,
@@ -502,7 +502,7 @@ find_formula.merMod <- function(x, ...) {
 
   f.cond <- stats::as.formula(get_fixed_effects(f.cond))
 
-  compact_list(list(conditional = f.cond, random = f.random))
+  .compact_list(list(conditional = f.cond, random = f.random))
 }
 
 
@@ -525,7 +525,7 @@ find_formula.rlmerMod <- function(x, ...) {
 
   f.cond <- stats::as.formula(get_fixed_effects(f.cond))
 
-  compact_list(list(conditional = f.cond, random = f.random))
+  .compact_list(list(conditional = f.cond, random = f.random))
 }
 
 
@@ -548,7 +548,7 @@ find_formula.mixed <- function(x, ...) {
 
   f.cond <- stats::as.formula(get_fixed_effects(f.cond))
 
-  compact_list(list(conditional = f.cond, random = f.random))
+  .compact_list(list(conditional = f.cond, random = f.random))
 }
 
 
@@ -571,7 +571,7 @@ find_formula.clmm <- function(x, ...) {
 
   f.cond <- stats::as.formula(get_fixed_effects(f.cond))
 
[…]
