### 2019-06-05 d8fbff5
mark internal functions
