### 2019-06-05 f2715cd
mark internal functions
