+++ R/get_parameters.R
@@ -42,7 +42,7 @@ get_parameters <- function(x, ...) {
 
 #' @export
 get_parameters.default <- function(x, ...) {
-  if (inherits(x, "list") && obj_has_name(x, "gam")) {
+  if (inherits(x, "list") && .obj_has_name(x, "gam")) {
     x <- x$gam
     class(x) <- c(class(x), c("glm", "lm"))
     return(get_parameters.gam(x, ...))
