+++ paper/paper.md
@@ -174,4 +174,6 @@ A last example is the **performance**-package [@pkg_performance], which provides
 
 *insight* is licensed under the GNU General Public License (v3.0), with all source code stored at GitHub (https://github.com/easystats/insight), with a corresponding issue tracker for bug-reporting and feature enhancements. In the spirit of open science and research, we encourage interaction with our package through requests/tips for fixes, support for additional model objects, feature updates, as well as general questions and concerns via direct interaction with contributors and developers.
 
+\pagebreak
+
 # References
