+++ paper/paper.bib
@@ -8,7 +8,7 @@
 }
 
 @Manual{pkg_performance,
-  title = {performance: Assessment of Regression Models Performance},
+  title = {{performance}: Assessment of Regression Models Performance},
   author = {Daniel Lüdecke and Dominique Makowski and Philip Waggoner},
   year = {2019},
   note = {R package version 0.2.0},
@@ -16,7 +16,7 @@
 }
 
 @Manual{pkg_sjPlot,
-  title = {sjPlot: Data Visualization for Statistics in Social Science},
+  title = {{sjPlot}: Data Visualization for Statistics in Social Science},
   author = {Daniel Lüdecke},
   year = {2019},
   note = {R package version 2.6.3},
@@ -35,7 +35,7 @@
 }
 
 @article{ludecke_ggeffects_2018,
-	title = {ggeffects: Tidy Data Frames of Marginal Effects from Regression Models},
+	title = {{ggeffects}: Tidy Data Frames of Marginal Effects from Regression Models},
 	volume = {3},
 	issn = {2475-9066},
 	url = {http://joss.theoj.org/papers/10.21105/joss.00772},
