+++ paper/paper.bib
@@ -7,24 +7,26 @@
   url = {https://www.R-project.org/},
 }
 
-@Manual{pkg_performance,
+@article{pkg_performance,
   title = {{performance}: Assessment of Regression Models Performance},
   author = {Daniel Lüdecke and Dominique Makowski and Philip Waggoner},
+  journal = {CRAN},
   year = {2019},
   note = {R package version 0.2.0},
   url = {https://easystats.github.io/performance/},
 }
 
-@Manual{pkg_sjPlot,
+@article{pkg_sjPlot,
   title = {{sjPlot}: Data Visualization for Statistics in Social Science},
   author = {Daniel Lüdecke},
+  journal = {CRAN},
   year = {2019},
   note = {R package version 2.6.3},
   url = {https://CRAN.R-project.org/package=sjPlot},
   doi = {10.5281/zenodo.1308157},
 }
 
-@Article{pkg_bayestestR,
+@article{pkg_bayestestR,
   title = {Understand and Describe Bayesian Models and Posterior Distributions using bayestestR},
   author = {Dominique Makowski and Mattan S. Ben-Shachar and Daniel Lüdecke},
   journal = {CRAN},
