### 2019-06-05 1d10524
Many grammatical fixes throughout
