+++ paper/paper.md
@@ -49,21 +49,21 @@ In total, the *insight* package includes 16 core functions (see Figure 1): `get_
 
 ## Definition of Model Components
 
-The functions from *insight* address different components of a model, however, due to conceptional overlap, there might be confusion about the specific "targets" of each function. We now provide a short explanation how *insight* defines components of regression models (see Figures 2 and 3). For detailed examples, we point to the [accompanying website](https://easystats.github.io/insight/articles/insight.html).
+The functions from *insight* address different components of a model. In an effort to avoid confusion about specific "targets" of each function, in this section we provide a short explanation of *insight*'s definitions of regression model components (see Figures 2 and 3). For detailed examples, we point users to the [accompanying package website](https://easystats.github.io/insight/articles/insight.html).
 
-* **data**: the dataset used to fit the model
-* **response**: the outcome or response variable (dependent variable) of a regression model
-* **predictor**: independent variables of (the _fixed_ part of) a regression model. For mixed models, variables that are (only) in the _random effects_ part (i.e. grouping factors) of the model are not returned as predictors by default, however, these can be included using additional arguments to the function call; predictors are "unqiue", hence if a variable appears as fixed effect and random slope, it is considered as one predictor (it is the same variable)
-* **random slopes**: variables that are used as random slope in a mixed effects model
-* **random or grouping factors**: variables that are used as grouping variables in a mixed effects model
-* **parameters**: values estimated or learned from data that encapsulate the relationship between variables; in regressions, these are usually referred to as *coefficients*
+* **data**: the dataset used to fit the model.
+* **response**: the outcome or response variable (dependent variable) of a regression model.
+* **predictor**: independent variables of (the _fixed_ part of) a regression model. For mixed models, variables that are only in the _random effects_ part (i.e. grouping factors) of the model are not returned as predictors by default. However, these can be included using additional arguments in the function call, treating predictors are "unqiue." As such, if a variable appears as a fixed effect and a random slope, it is treated as one (the same) predictor.
+* **random slopes**: variables that are specified as random slopes in a mixed effects model.
+* **random or grouping factors**: variables that are specified as grouping variables in a mixed effects model.
+* **parameters**: values estimated or learned from data that capture the relationship between variables. In regression models, these are usually referred to as *coefficients*.
 
 ![Definition of Model Components, Part 1](figure2a.png)
 
 \pagebreak
 
-* **term**: terms are any (unique) variables that appear in a regression model, like response variables, predictors or random effects; a "term" only relates to the unique occurence of a variable; for instance, in the expression `x + I(x^2)`, there is only the term `x`
-* **variables**: a variable is considered as an object that stores unique data information; for instance, the expression `x + I(x^2)` has two objects with two different sets of data values, and thus are treated as two variables
+* **term**: any unique variables that appear in a regression model, e.g., response variables, predictors or random effects. A "term" only relates to the unique occurence of a variable. For instance, the expression `x + I(x^2)` has only the term `x`.
+* **variables**: an object that stores unique data information. For instance, the expression `x + I(x^2)` has two objects with two different sets of data values, and thus are treated as two variables.
 
 ![Definition of Model Components, Part 2](figure2b.png)
 
@@ -158,13 +158,13 @@ find_formula(sample3)
 
 **insight** is already used by different packages to solve problems that typically occur when the users' inputs are different model objects of varying complexity.
 
[…]
