### 2019-06-05 #1412 @strengejacke
Dear @alexpghayes,

on behalf of the co-authors, let me thank you again for your thorough comments on our submission. We substantially revised our paper, the package documentation and the accompanying website, and revised package code and functions as suggested. Let me give you a more detailed response to your suggestions and our revisions.

> Interface complexity: My primary concern is that insight has a complicated interface. With such a broad interface, I had a hard time figuring out the entry point. Where should I start?

Thank you for this comment. We have revised the paper accordingly, making it more clear for users where they can start. We adapted these revisions both to README of the package’s front page (https://github.com/easystats/insight), as well as the online documentation (https://easystats.github.io/insight/articles/insight.html).

> What is the difference between get_*() and find_*()?

We have added more details in the paper, to the README and the online documentation to make this essential distinction clearer to users. Furthermore, in addition with the figure showing how package functions are related to model components (Figure 1 in the paper), it should now be easier to grasp the differences between the get- and the find-functions.

> I strongly recommend shrinking this interface.

See our previous point. In revising the paper and the online documentation, the need for both get- and find-functions should be clearer now. Since it is not possible to extract parameter names for specific model parts (like random or fixed effects, zero-inflated or conditional component) without additional effort when you have the data from one of the get-functions, and vice versa, getting the underlying data once you have the names of parameters or predictors, we see the need for having both find- and get-functions. There are several use cases where either one is needed, and getting the required information or data easily is one of the core aspects of the insight package.

> Also, I struggled to understand the difference between: variables and predictors and response, terms and formula. This meant I had to read the documentation for each of these methods before figuring out which one did what I wanted. These distinctions should be explained clearly on the front page of the pagedown website. The diagram in the JOSS paper helps somewhat but still needs an accompanying narrative.

Thanks for pointing this out. We have added a complete new section on the definition of model components to the paper. Furthermore, we added two figures that help making it more clear what the differences between these terms are, and why it is necessary to have these distinctions. We also added the new section to the README of the package’s front page, as well as to the online-documentation (including the new figures).

> While the documentation is very clear about function returns, I had difficulty understanding why behavior differs from different model objects. In find_parameters(): aov objects fit linear models with conditional and random terms, but instead this information is stored in between and within fields of the return

We agree with this point, and revised affected package functions accordingly. Now, there are less exceptions, making it more clear and expectable for users what will be returned. Regarding your particular example, the functions now return the elements “conditional” and “random”, being in line with the return values of other model objects.

> In find_algorithm(): the object return differs from different object types, so I still would need to check which elements are in this list before trying to access them. Also, I belive the term algorithm should be replaced by estimator, which is more appropriate. Also, for Bayesian models, describing the algorithm as "sampling" is too vague.

We agree with this point. We implemented this function, knowing that it currently is still a bit “work in progress”, and there is an open GitHub issue on this (https://github.com/easystats/insight/issues/38). So this is something we plan to address, however, we still need and want to ask other people with correspondent expertise for their opinion in order to adequately address this issue.

> In find_formula(): how is random different from correlation? How is slopes different from conditional? I would have grouped these together.

This is a design decision where we think it is not easily possible to find a closing decision. Our idea was to reflect the different parts of a model formula in the returned elements from the function. “lme()”, for instance, can have both a random-argument and a correlation-argument, and we wanted to preserve this distinction. For “feis”-objects (fixed effects regressions with individual slopes), we have a complicated formula structure with several components. Again, we wanted to preserve the distinction between these parts of the model formula. However, we are happy to reduce interface complexity here and are open for concrete suggestions that help users and developers.

> Examples: The documentation shows how a user can extract information from a model, but it doesn't show how the user can use this information to solve larger problems. Some applications showing how the various functions operate together (and how they are distinct) would be greatly beneficial.

Thanks for this suggestion. We have added a new section with examples of practical use cases (in different R packages) to our paper.

> /R is organized by method but /test is organized by model

This is due to performance-reasons. Since each function should work for (almost) every model object, for each test we would need to fit all the models each time again, if tests were organized by models. To reduce computational time, especially for CRAN, we decided to organize tests differently that the code-files in the R-folder.

> The homepage and README should have links to information about contributing

We have added explicit information about information, support and contributing to the README and website.

> I think the package needs narrative explaining at a conceptual level how it generalizes across models, and return objects with fewer special cases.

We have largely revised the paper, README and website (online documentation), providing a clearer narrative (also at the conceptual level) in order to make the idea and functioning of the package clearer for users.

> insight may benefit from narrowing in on particular families of models. For instance, it seems like most existing functionality is targeted at GLMMs.

One of the core aims of the insight package is to provide a unified syntax to access model information. Since this information is not easily accessible for many model objects, the idea is to work with as many models as possible. Against the background, it would run against the nature of this package to narrow the scope of supported models. We have added a section in the paper showing examples of use cases and hope that these examples explain why it makes sense to support many different models.

Again, thank your for this valuable review! We hope that we have addressed all your points sufficiently and are happy to receive your feedback on our revision.

Best
Daniel, on behalf of the co-authors
