### 2019-06-05 #1412 @strengejacke
If the paper.md does not yet compile, we probably can use a temporary PDF?
Here is the link: https://github.com/easystats/insight/blob/master/paper/paper.pdf
