### 2019-06-05 14c706a
add support for survfit
