+++ NAMESPACE
@@ -131,6 +131,7 @@ S3method(get_data,rlmerMod)
 S3method(get_data,rqss)
 S3method(get_data,stanmvreg)
 S3method(get_data,stanreg)
+S3method(get_data,survfit)
 S3method(get_data,tobit)
 S3method(get_data,vgam)
 S3method(get_data,vglm)
@@ -223,6 +224,7 @@ S3method(link_function,rq)
 S3method(link_function,speedglm)
 S3method(link_function,speedlm)
 S3method(link_function,stanmvreg)
+S3method(link_function,survfit)
 S3method(link_function,survreg)
 S3method(link_function,svyolr)
 S3method(link_function,tobit)
@@ -282,6 +284,7 @@ S3method(link_inverse,rq)
 S3method(link_inverse,speedglm)
 S3method(link_inverse,speedlm)
 S3method(link_inverse,stanmvreg)
+S3method(link_inverse,survfit)
 S3method(link_inverse,survreg)
 S3method(link_inverse,svyolr)
 S3method(link_inverse,tobit)
@@ -341,6 +344,7 @@ S3method(model_info,rqss)
 S3method(model_info,speedglm)
 S3method(model_info,speedlm)
 S3method(model_info,stanmvreg)
+S3method(model_info,survfit)
 S3method(model_info,survreg)
 S3method(model_info,svyolr)
 S3method(model_info,tobit)
@@ -372,6 +376,7 @@ S3method(n_obs,mlogit)
 S3method(n_obs,multinom)
 S3method(n_obs,rq)
 S3method(n_obs,stanmvreg)
+S3method(n_obs,survfit)
 S3method(n_obs,svyolr)
 S3method(n_obs,wbm)
 S3method(n_obs,zeroinfl)
