### 2019-06-10 fd00345
adding philip as aut/ctb
forgot to add myself after the code updates a while back; now rectified
