### 2019-06-10 71f847d
fix missing weights-col
