### 2019-06-10 fb9f5fc
add find and get weights
