+++ NAMESPACE
@@ -395,6 +395,7 @@ export(find_random_slopes)
 export(find_response)
 export(find_terms)
 export(find_variables)
+export(find_weights)
 export(get_correlation_slope_intercept)
 export(get_data)
 export(get_parameters)
@@ -410,6 +411,7 @@ export(get_variance_intercept)
 export(get_variance_random)
 export(get_variance_residual)
 export(get_variance_slope)
+export(get_weights)
 export(has_intercept)
 export(is_model)
 export(is_multivariate)
