### 2019-06-10 d8a0e2f
try three sources to find weights
