+++ R/find_weights.R
@@ -10,7 +10,8 @@
 #'
 #' @examples
 #' data(mtcars)
-#' m <- lm(mpg ~ wt + cyl + vs, data = mtcars)
+#' mtcars$weight <- rnorm(nrow(mtcars), 1, .3)
+#' m <- lm(mpg ~ wt + cyl + vs, data = mtcars, weights = weight)
 #' find_weights(m)
 #' @export
 find_weights <- function(x) {
