+++ man/find_weights.Rd
@@ -18,6 +18,7 @@ Returns the name of the variable that describes the weights of a model.
 }
 \examples{
 data(mtcars)
-m <- lm(mpg ~ wt + cyl + vs, data = mtcars)
+mtcars$weight <- rnorm(nrow(mtcars), 1, .3)
+m <- lm(mpg ~ wt + cyl + vs, data = mtcars, weights = weight)
 find_weights(m)
 }
