+++ man/get_weights.Rd
@@ -17,6 +17,8 @@ Returns weighting variable of a model.
 }
 \examples{
 data(mtcars)
-m <- lm(mpg ~ wt + cyl + vs, data = mtcars)
-find_weights(m)
+mtcars$weight <- rnorm(nrow(mtcars), 1, .3)
+m <- lm(mpg ~ wt + cyl + vs, data = mtcars, weights = weight)
+get_weights(m)
+
 }
