### 2019-06-11 fd4fa48
test find/get weight
