+++ tests/testthat/test-negbin.R
@@ -6,7 +6,7 @@ if (.runThisTest && Sys.getenv("USER") != "travis") {
     context("insight, negbin")
 
     data(dja)
-    m1 <- negbin(y ~ group + offset(log(trisk)), ~ village, data = dja)
+    m1 <- aod::negbin(y ~ group + offset(log(trisk)), random = ~ village, data = dja)
 
     test_that("model_info", {
       expect_true(model_info(m1)$is_negbin)
