+++ DESCRIPTION
@@ -9,7 +9,7 @@ Description: A tool to provide an easy, intuitive and consistent access to
    'find_', and functions that get the underlying data, starting with 'get_'. 
    The package has a consistent syntax and works with many different model 
    objects, where otherwise functions to access these information are missing.
-Version: 0.3.1
+Version: 0.3.0.9000
 Date: 2019-06-05
 Authors@R: c(person("Daniel", "Lüdecke", role = c("aut", "cre"), email = "d.luedecke@uke.de", comment = c(ORCID = "0000-0002-8895-3206")),
              person("Dominique", "Makowski", role = c("aut", "ctb"), email = "dom.makowski@gmail.com", comment = c(ORCID = "0000-0001-5375-9967")),
