+++ tests/testthat/test-rstanarm.R
@@ -13,12 +13,17 @@ if (.runThisTest || Sys.getenv("USER") == "travis") {
     m2 <- insight::download_model("stanreg_glm_6")
     m3 <- insight::download_model("stanreg_glm_1")
 
+    m4 <- stan_glm(RT ~ color * shape, data = puzzles, prior = cauchy(0, c(3, 1, 2)), iter = 500, chains = 2)
+    m5 <- stan_glm(RT ~ color * shape, data = puzzles, prior = cauchy(0, c(1, 2, 3)), iter = 500, chains = 2)
+
     test_that("get_priors", {
       expect_equal(colnames(get_priors(m1)), c("parameter", "distribution", "location", "scale"))
       expect_equal(colnames(get_priors(m2)), c("parameter", "distribution", "location", "scale", "adjusted_scale"))
       expect_equal(get_priors(m1)$scale, c(10.0, 2.5, 2.5, 2.5, 2.5), tolerance = 1e-3)
-      expect_equal(get_priors(m2)$adjusted_scale, c(4.3586628, 0.4119705, 0.5360283, 0.6172700, 1.0896657, 1.0896657), tolerance = 1e-3)
+      expect_equal(get_priors(m2)$adjusted_scale, c(4.35866284936698, 1.08966571234175, 1.08966571234175, 0.617270040728345, 0.536028320794122, 0.411970489526739), tolerance = 1e-3)
       expect_equal(get_priors(m3)$adjusted_scale, c(NA, 2.555042), tolerance = 1e-3)
+      expect_equal(get_priors(m4)$adjusted_scale, c(25.5992021152256, 7.67976063456768, 2.55992021152256, 5.11984042304512), tolerance = 1e-3)
+      expect_equal(get_priors(m5)$adjusted_scale, c(25.5992021152256, 2.55992021152256, 5.11984042304512, 7.67976063456768), tolerance = 1e-3)
     })
 
 
