### 2019-06-15 e71ff29
safer data access from environment
