### 2019-06-16 7254f9b
fix test issues
