+++ man/clean_names.Rd
@@ -20,8 +20,8 @@ This function "cleans" names of model terms (or a character
   \code{as.factor()} etc.
 }
 \note{
-If \code{x} is a regression model, this function is equal to calling
-  \code{find_terms()}.
+If \code{x} is a regression model, this function is (almost) equal to
+  calling \code{find_terms()}.
 }
 \examples{
 # example from ?stats::glm
