### 2019-06-16 #1412 @alexpghayes
## Response to revisions

Response based on comments by @strengejacke above and the Github master branch on 2019-06-16.

I'm impressed by the changes you've made, especially to the documentation. After reading more carefully, I agree with your comments that the broad interface is warranted.

Something clicked for me when I realized that the importance of differentiating between names and values. Some functions return the names of columns of data, some functions return values of columns of data. The difference is especially clear with `get_weights()` and `find_weights()`. I believe `get_*()` and `find_*()` are an attempt to make this differentiation, but I am now convinced that a better approach is to give functions names like the following:

- `find_weights()` -> `weights_name()`
- `get_weights()` -> `weights_values()`

I suspect users will have an easier time with this sort of naming strategy and encourage you to consider this scheme, or something similar.

## Documentation

I appreciated the new definitions of the various components of a model. Instead of presenting this material all at once, I recommend you group the definitions by components that are easy to confuse and set these groups visually apart.

Note that you have chosen to define `terms` and `variables` in a non-standard way for the R community. In an R model formula an entire chunk such as `I(x^2)` is referred to as a term, rather than a variable. I recommend following this definition, which amounts to flipping the current usage of *term* and *variable*.

Finally, I recommend grouping `get_*()` and `find_*()` together on the pkgdown site in the Reference section. That is, you want `get_parameters()` and `find_parameters()` (or `parameter_names()` and `parameter_values()`) to appear next to each other.

### Miscellenea

- I appreciate that there are fewer special cases in the returns.

- The test file organization makes sense, I'm on board.

- It makes me nervous that the example at https://easystats.github.io/insight/reference/clean_parameters.html warns about overwriting quosure methods. Can you comment on why this is happening and/or necessary?

- When is `clean_names()` different from `find_terms()`? I would include examples so that users can tell when to use which function.

- I appreciate that you've linked to packages that use `insight`. However, this is not especially useful to users -- instead of demonstrating how to solve a specific problem with `insight`, the user must now go read an entire package on their own, find a usage of `insight`, and infer the problem that is being solved. I would construct explicit, self-contained examples showing idiomatic usage. If these are inspired by the linked packages, fantastic, but merely linking to another code base isn't that helpful.

- I recommend keeping the original paper title `insight: Easy Access to Model Information for Various Model Objects` rather than `insight: A Unified Syntax for Accessing Model Information for Various Model Objects in R`. Insight provides a unified *interface*, but does not introduce any new *syntax*, which is a language level feature. Domain specific languages such as `dplyr` and `data.table` introduce new syntax (i.e. `:=` and `!!`) via metaprogramming tricks, but `insight` does not and makes use of standard evaluation and syntax.

- I would rename `all_models_equal()` to `all_models_have_same_class()`. The current name makes it seem like the `all_models_equal()` is also checking whether the parameter values match up.

- I am about to respond with additional thoughts on `find_algorithm()` in https://github.com/easystats/insight/issues/38.
