### 2019-06-17 69907ca
draft combine parameter objects
