+++ NAMESPACE
@@ -384,6 +384,7 @@ S3method(n_obs,zerotrunc)
 export(all_models_equal)
 export(clean_names)
 export(clean_parameters)
+export(combine_parameter_objects)
 export(download_model)
 export(find_algorithm)
 export(find_formula)

[…]
