### 2019-06-17 #1412 @strengejacke
@alexpghayes Thanks for your review! Since some of your remaining suggestions mean quite some substantial rework or breaking changes, I would like to clarify some points and suggest how we would address your issues, before we actually start working on it. Could you please give a short feedback if our planned steps are ok for you?

1. Rename find_weights() -> weights_name() and get_weights() -> weights_values()

Since _insight_ is part of the larger "easystats" project, we have developed some coding guidelines for our packages (https://github.com/easystats/easystats#convention-of-code-style). Your suggestion would break these guidelines, so for consistency within and across packages, we would prefer to keep the current names. However, we may add aliases to the existing functions.

2. Definition of "term" and "variable".

I'm not sure about this, I could not find clear definitions yet (if you have some references/links, would be happy if you could share them). However, since these functions are currently only used by one package that depends on _insight_, we could flip the usage of _term_ and _variable_ here.

3. Finally, I recommend grouping get_*() and find_*() together on the pkgdown site in the Reference section.

I'm not sure if this is possible with some _pkgdown_ options, I think we that have to combine the functions rd-files, so both `?find_parameters` and `?get_parameters` would appear in the same help-page. Would you suggest going this route?

4. It makes me nervous that the example at https://easystats.github.io/insight/reference/clean_parameters.html warns about overwriting quosure methods. Can you comment on why this is happening and/or necessary?

This seems to be new in R 3.6 (at least, since then it started appearing for me), and once namespaces are referred to, R warns about overwritten S3-methods. ggplot2 depends on rlang, afaik, and it seems ggplot2 overwrites some S3-methods from rlang (or vice versa). In short, this issues seems unrelated to _insight_.

5. I appreciate that you've linked to packages that use insight. However, this is not especially useful to users -- instead of demonstrating how to solve a specific problem with insight, the user must now go read an entire package on their own, find a usage of insight, and infer the problem that is being solved.

We would add some more comprehensive working examples that clearly show some solutions to problems to the readme (and as further (sub) section to the paper) - would that be sufficient?
