### 2019-06-17 c9f0a5c
combine_parameter_objects -> print_multiple_parameters
