+++ man/print_multiple_parameters.Rd
@@ -1,11 +1,11 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/combine_parameter_objects.R
-\name{combine_parameter_objects}
-\alias{combine_parameter_objects}
+% Please edit documentation in R/print_multiple_parameters.R
+\name{print_multiple_parameters}
+\alias{print_multiple_parameters}
 \title{Get clean names of model paramters}
 \usage{
-combine_parameter_objects(x, ..., split_by = c("Effects", "Component",
-  "Response"))
+print_multiple_parameters(x, ..., split_by = c("Effects", "Component",
+  "Group", "Response"))
 }
 \arguments{
 \item{x}{A fitted model.}
@@ -31,6 +31,6 @@ library(bayestestR)
 model <- download_model("brms_zi_2")
 x <- hdi(model, effects = "all", component = "all")
 x
-combine_parameter_objects(model, x)
+print_multiple_parameters(model, x)
 
 }
