+++ NAMESPACE
@@ -384,7 +384,6 @@ S3method(n_obs,zerotrunc)
 export(all_models_equal)
 export(clean_names)
 export(clean_parameters)
-export(combine_parameter_objects)
 export(download_model)
 export(find_algorithm)
 export(find_formula)
@@ -423,6 +422,7 @@ export(model_info)
 export(n_obs)
 export(print_color)
 export(print_colour)
+export(print_multiple_parameters)
 importFrom(stats,Gamma)
 importFrom(stats,as.formula)
 importFrom(stats,binomial)
