### 2019-06-18 8839280
add issue-nr to news
