### 2019-06-18 44bd133
add docs to print_parameters
