+++ R/print_parameters.R
@@ -1,0 +1,167 @@
+#' @title Prepare summary statistics of model paramters for printing
+#' @name print_parameters
+#'
+#' @description This function takes a data frame, typically a data frame with
+#' information on summaries of model parameters like \code{\link[bayestestR]{hdi}}
+#' or \code{\link[bayestestR]{equivalence_test}}, as input and splits this information
+#' into several parts, depending on the model. See details below.
+#'
+#' @param x A fitted model, or a data frame returned by \code{\link{clean_parameters}}.
+#' @param ... One or more objects (data frames), which contain information about
+#'   the model parameters and related statistics (like confidence intervals, HDI,
+#'   ROPE, ...).
+#' @param split_by If \code{NULL}, the merged data frame is returned. Else, the
+#'   data frame is split into a list, split by the values from those columns defined
+#'   in \code{split_by}. \code{split_by} should be a character vector with one or
+#'   more of the following elements: \code{"Effects"}, \code{"Component"},
+#'   \code{"Response"} and \code{"Group"}. These are the column names returned
+#'   by \code{\link{clean_parameters}}, which is used to extract the information
+#'   to which group or component model parameters belong.
+#'
+#' @return A data frame or a list of data frame (if \code{split_by} is not \code{NULL}).
+#' If a list is returned, the element names reflect the model components where the
+#' extracted information in the data frames belong to, e.g. \code{`random.zero_inflated.Intercept: persons`}
+#' is the data frame that contains the parameters for the random effects from
+#' group-level "persons" from the zero-inflated model component.
+#'
+#' @details This function is intended to prepare data frames that contain information
+#' about model parameters for clear printing.
+#' \cr \cr
+#' First, \code{x} is required, which should either be a model object or a
+#' prepared data frame as returned by \code{\link{clean_parameters}}. If
+#' \code{x} is a model, \code{clean_parameters()} is called on that model
+#' object to get information with which model components the parameters
+#' are associated.
+#' \cr \cr
+#' Then, \code{...} take one or more data frames that also contain information
+#' about parameters from the same model, but also have additional information
+#' provided by other methods. For instance, a data frame in \code{...} might
+#' be the result of \code{\link[bayestestR]{hdi}}, where we
+#' have a) a \code{Parameters} column and b) columns with the HDI values.
+#' \cr \cr
+#' Now we have a data frame with model parameters and information about the
+#' association to the different model components, and a data frame with model
+#' parameters and some summary statistics. \code{print_parameters()}
+#' then merges these data frame, so the statistic of interest (in our example:
+#' the HDI) is also associated with the different model components. The data
+#' frame is split into a list, so for a clearer printing one can loop over this
+#' list and print each component for a better overview. Furthermore, parameter
+#' names are "cleaned", if necessary, also for a cleaner print. See also 'Examples'.
+#'
+#' @examples
+#' library(bayestestR)
+#' model <- download_model("brms_zi_2")
+#' x <- hdi(model, effects = "all", component = "all")
+#'
+#' # hdi() returns a data frame, here we use only the informaton on
+#' # parameter names and HDI values
+#' tmp <- as.data.frame(x)[, 1:4]
+#' tmp
+#'
+#' # Based on the "split_by" argument, we get a list of data frames that
+#' # is split into several parts that reflect the model components.
[…]
