+++ R/clean_parameters.R
@@ -10,7 +10,7 @@
 #' in particular of - but not limited to - posterior samples from Stan models,
 #' depending on certain characteristics. This might be useful when only selective
 #' results should be reported or results from all parameters should be filtered
-#' to return only certain results.
+#' to return only certain results (see \code{\link{print_parameters}}).
 #'
 #' @param x A fitted model.
 #' @inheritParams find_parameters
