+++ NAMESPACE
@@ -422,7 +422,7 @@ export(model_info)
 export(n_obs)
 export(print_color)
 export(print_colour)
-export(print_multiple_parameters)
+export(print_parameters)
 importFrom(stats,Gamma)
 importFrom(stats,as.formula)
 importFrom(stats,binomial)

[…]
