### 2019-06-18 fe217b3
update news
