### 2019-06-18 ea60d5d
include intercept in polr find/get parameters
