+++ NAMESPACE
@@ -91,6 +91,7 @@ S3method(find_parameters,lme)
 S3method(find_parameters,lrm)
 S3method(find_parameters,merMod)
 S3method(find_parameters,mixed)
+S3method(find_parameters,polr)
 S3method(find_parameters,rlmerMod)
 S3method(find_parameters,rqss)
 S3method(find_parameters,stanmvreg)
@@ -162,6 +163,7 @@ S3method(get_parameters,lme)
 S3method(get_parameters,lrm)
 S3method(get_parameters,merMod)
 S3method(get_parameters,mixed)
+S3method(get_parameters,polr)
 S3method(get_parameters,rlmerMod)
 S3method(get_parameters,rqss)
 S3method(get_parameters,stanmvreg)
