+++ NEWS.md
@@ -19,6 +19,7 @@
 ## Changes to functions
 
 * `find_parameters()` and `get_parameters()` now allow to return a `sigma`-element for multivariate-response models (*brmsfit*, *stanmvreg*).
+* `find_parameters()` and `get_parameters()` now return the intercepts for **polr** models.
 * `find_parameters()` gets an `effects` and a `component`-argument, similar to many other functions in _insight_.
 * `get_priors()` now returns `distribution = "uniform"` when model was fitted with flat priors.
 * `model_info()` now returns an element `$is_hurdle` for hurdle models.
