### 2019-06-18 b69f85a
JOSS: revert back to old paper title
