### 2019-06-18 18e3cdc
handle prior and simplex parameter
