### 2019-06-19 199e2f4
fix test issues
