### 2019-06-19 79d6437
fix test issues
