### 2019-06-19 #1412 @alexpghayes
> Since insight is part of the larger "easystats" project, we have developed some coding guidelines for our packages (https://github.com/easystats/easystats#convention-of-code-style). Your suggestion would break these guidelines, so for consistency within and across packages, we would prefer to keep the current names. However, we may add aliases to the existing functions.

If you don't want to change these, I understand. Since the difference between `get_()` and `find_()` does seem to come down to names versus values, I really do recommend function names reflecting this, but I'm happy to recommend for publication either way.

> Definition of "term" and "variable".

Definitely swap. It look like you found some references in https://github.com/easystats/insight/issues/56#issuecomment-503565591?

> Finally, I recommend grouping get_() and find_() together on the pkgdown site in the Reference section.

Yeah, you can do this all with `pkgdown` options, no need for `@rdname` or anything fancy. See the `_pkgdown.yaml` file (I think that's what it's called) in `dplyr` for examples: https://dplyr.tidyverse.org/reference/index.html

> This seems to be new in R 3.6 (at least, since then it started appearing for me), and once namespaces are referred to, R warns about overwritten S3-methods. ggplot2 depends on rlang, afaik, and it seems ggplot2 overwrites some S3-methods from rlang (or vice versa). In short, this issues seems unrelated to insight.

Great, wouldn't worry about it then.

> We would add some more comprehensive working examples that clearly show some solutions to problems to the readme (and as further (sub) section to the paper) - would that be sufficient?

Yep, that'd be ideal! The current examples show how to access certain kinds of information. The open question is *when* would users want to access this kind of information. What common problem does `insight` provide a solution too? My guess is there's some problem where you need to map over many models at once for model comparison and `insight` makes that process nicer.
