### 2019-06-19 4e7759c
flip find_variables <> find_terms
