+++ tests/testthat/test-speedlm.R
@@ -63,10 +63,10 @@ if (require("testthat") && require("insight") && require("speedglm")) {
     )
   })
 
-  test_that("find_terms", {
-    expect_equal(find_terms(m1), list(response = "Sepal.Length", conditional = c("Petal.Width", "Species")))
-    expect_equal(find_terms(m1, flatten = TRUE), c("Sepal.Length", "Petal.Width", "Species"))
-    expect_equal(find_terms(m2, flatten = TRUE), c("mpg", "hp", "cyl", "wt"))
+  test_that("find_variables", {
+    expect_equal(find_variables(m1), list(response = "Sepal.Length", conditional = c("Petal.Width", "Species")))
+    expect_equal(find_variables(m1, flatten = TRUE), c("Sepal.Length", "Petal.Width", "Species"))
+    expect_equal(find_variables(m2, flatten = TRUE), c("mpg", "hp", "cyl", "wt"))
   })
 
   test_that("n_obs", {
@@ -95,16 +95,16 @@ if (require("testthat") && require("insight") && require("speedglm")) {
     expect_false(is_multivariate(m1))
   })
 
-  test_that("find_variables", {
+  test_that("find_terms", {
     expect_equal(
-      find_variables(m1),
+      find_terms(m1),
       list(
         response = "Sepal.Length",
         conditional = c("Petal.Width", "Species")
       )
     )
     expect_equal(
-      find_variables(m2),
+      find_terms(m2),
       list(
         response = "log(mpg)",
         conditional = c("log(hp)", "cyl", "I(cyl^2)", "poly(wt, degree = 2, raw = TRUE)")
