+++ tests/testthat/test-gbm.R
@@ -62,9 +62,9 @@ if (.runThisTest || Sys.getenv("USER") == "travis") {
       )
     })
 
-    test_that("find_terms", {
-      expect_equal(find_terms(m1), list(response = "mpg", conditional = c("gear", "cyl", "wt")))
-      expect_equal(find_terms(m1, flatten = TRUE), c("mpg", "gear", "cyl", "wt"))
+    test_that("find_variables", {
+      expect_equal(find_variables(m1), list(response = "mpg", conditional = c("gear", "cyl", "wt")))
+      expect_equal(find_variables(m1, flatten = TRUE), c("mpg", "gear", "cyl", "wt"))
     })
 
     test_that("n_obs", {
@@ -89,9 +89,9 @@ if (.runThisTest || Sys.getenv("USER") == "travis") {
       expect_equal(get_parameters(m1)$parameter, c("wt", "gear", "cyl"))
     })
 
-    test_that("find_variables", {
+    test_that("find_terms", {
       expect_equal(
-        find_variables(m1),
+        find_terms(m1),
         list(
           response = "mpg",
           conditional = c("gear", "cyl", "wt")
