+++ tests/testthat/test-gam.R
@@ -43,9 +43,9 @@ if (require("testthat") && require("insight") && require("mgcv")) {
     )
   })
 
-  test_that("find_terms", {
-    expect_equal(find_terms(m1), list(response = "y", conditional = c("x0", "x1", "x2", "x3")))
-    expect_equal(find_terms(m1, flatten = TRUE), c("y", "x0", "x1", "x2", "x3"))
+  test_that("find_variables", {
+    expect_equal(find_variables(m1), list(response = "y", conditional = c("x0", "x1", "x2", "x3")))
+    expect_equal(find_variables(m1, flatten = TRUE), c("y", "x0", "x1", "x2", "x3"))
   })
 
   test_that("n_obs", {
@@ -75,9 +75,9 @@ if (require("testthat") && require("insight") && require("mgcv")) {
     expect_false(is_multivariate(m1))
   })
 
-  test_that("find_variables", {
+  test_that("find_terms", {
     expect_equal(
-      find_variables(m1),
+      find_terms(m1),
       list(
         response = "y",
         conditional = c("s(x0)", "s(x1)", "s(x2)", "s(x3)")
