+++ NAMESPACE
@@ -384,6 +384,7 @@ S3method(n_obs,wbm)
 S3method(n_obs,zeroinfl)
 S3method(n_obs,zerotrunc)
 export(all_models_equal)
+export(all_models_same_class)
 export(clean_names)
 export(clean_parameters)
 export(download_model)
