### 2019-06-20 2179dcb
add nobs() method
