+++ R/n_obs.R
@@ -11,6 +11,9 @@
 #' @return The number of observations used to fit the model, or \code{NULL} if
 #'   this information is not available.
 #'
+#' @note For model-objects supported by \pkg{insight} that \emph{do not} have
+#' a \code{nobs()}-method, \pkg{insight} provides a \code{nobs()}-method as well.
+#'
 #' @examples
 #' data(mtcars)
 #' m <- lm(mpg ~ wt + cyl + vs, data = mtcars)
@@ -40,6 +43,7 @@ n_obs.default <- function(x, ...) {
 }
 
 
+
 #' @rdname n_obs
 #' @export
 n_obs.svyolr <- function(x, weighted = FALSE, ...) {
@@ -50,6 +54,7 @@ n_obs.svyolr <- function(x, weighted = FALSE, ...) {
 }
 
 
+
 #' @export
 n_obs.gamm <- function(x, ...) {
   x <- x$gam
@@ -57,11 +62,23 @@ n_obs.gamm <- function(x, ...) {
   NextMethod()
 }
 
+#' @export
+nobs.gamm <- function(x, ...) {
+  x <- x$gam
+  class(x) <- c(class(x), c("glm", "lm"))
+  n_obs(x, ...)
+}
+
+
 
 #' @export
 n_obs.lmRob <- function(x, ...) {
   length(x$fitted.values)
 }
+
+#' @export
+nobs.lmRob <- n_obs.lmRob
+
 
 
 #' @export
@@ -75,11 +92,19 @@ n_obs.biglm <- function(x, ...) {
   x$n
 }
 
+#' @export
+nobs.biglm <- n_obs.biglm
+
+
 
 #' @export
 n_obs.bigglm <- function(x, ...) {
   x$n
 }
+
+#' @export
+nobs.bigglm <- n_obs.bigglm
+
 
 
 #' @export
@@ -87,11 +112,19 @@ n_obs.gbm <- function(x, ...) {
   length(x$fit)
 }
 
+#' @export
+nobs.gbm <- n_obs.gbm
+
+
 
 #' @export
 n_obs.glimML <- function(x, ...) {
   nrow(x@data)
 }
+
+#' @export
+nobs.glimML <- n_obs.glimML
+
 
 
 #' @export
@@ -99,11 +132,20 @@ n_obs.glmRob <- function(x, ...) {
   length(x$fitted.values)
 }
 
+#' @export
+nobs.glmRob <- n_obs.glmRob
+
+
 
 #' @export
 n_obs.gmnl <- function(x, ...) {
   x$logLik$nobs
 }
+
+#' @export
+nobs.gmnl <- n_obs.gmnl
+
+
 
 
 #' @export
@@ -111,11 +153,19 @@ n_obs.multinom <- function(x, ...) {
   nrow(x$fitted.values)
 }
 
+#' @export
+nobs.multinom <- n_obs.multinom
+
+
 
 #' @export
 n_obs.rq <- function(x, ...) {
   length(x$fitted.values)
 }
+
+#' @export
+nobs.rq <- n_obs.rq
+
 
 
 #' @export
