+++ NAMESPACE
@@ -383,6 +383,16 @@ S3method(n_obs,svyolr)
 S3method(n_obs,wbm)
 S3method(n_obs,zeroinfl)
 S3method(n_obs,zerotrunc)
+S3method(nobs,bigglm)
+S3method(nobs,biglm)
+S3method(nobs,gamm)
+S3method(nobs,gbm)
+S3method(nobs,glimML)
+S3method(nobs,glmRob)
+S3method(nobs,gmnl)
+S3method(nobs,lmRob)
+S3method(nobs,multinom)
+S3method(nobs,rq)
 export(all_models_equal)
 export(all_models_same_class)
 export(clean_names)
