+++ NAMESPACE
@@ -383,16 +383,30 @@ S3method(n_obs,svyolr)
 S3method(n_obs,wbm)
 S3method(n_obs,zeroinfl)
 S3method(n_obs,zerotrunc)
+S3method(nobs,BBmm)
+S3method(nobs,BBreg)
+S3method(nobs,aovlist)
 S3method(nobs,bigglm)
 S3method(nobs,biglm)
+S3method(nobs,coxme)
+S3method(nobs,coxph)
+S3method(nobs,crq)
+S3method(nobs,feis)
+S3method(nobs,felm)
 S3method(nobs,gamm)
 S3method(nobs,gbm)
 S3method(nobs,glimML)
 S3method(nobs,glmRob)
 S3method(nobs,gmnl)
+S3method(nobs,hurdle)
 S3method(nobs,lmRob)
+S3method(nobs,mlogit)
 S3method(nobs,multinom)
 S3method(nobs,rq)
+S3method(nobs,stanmvreg)
+S3method(nobs,survfit)
+S3method(nobs,zeroinfl)
+S3method(nobs,zerotrunc)
 export(all_models_equal)
 export(all_models_same_class)
 export(clean_names)
