### 2019-06-20 cd12859
demonstrate diff between cleannames and findvar
