+++ man/clean_names.Rd
@@ -20,8 +20,14 @@ This function "cleans" names of model terms (or a character
   \code{as.factor()} etc.
 }
 \note{
-If \code{x} is a regression model, this function is (almost) equal to
-  calling \code{find_variables()}.
+Typically, this method is intended to work on character vectors,
+  in order to remove patterns that obscure the variable names. For
+  convenience reasons it is also possible to call \code{clean_names()}
+  also on a model object. If \code{x} is a regression model, this
+  function is (almost) equal to calling \code{find_variables()}. The
+  main difference is that \code{clean_names()} always returns a character
+  vector, while \code{find_variables()} returns a list of character
+  vectors, unless \code{flatten = TRUE}. See 'Examples'.
 }
 \examples{
 # example from ?stats::glm
@@ -30,4 +36,17 @@ outcome <- c(gl(3, 1, 9))
 treatment <- gl(3, 3)
 m <- glm(counts ~ log(outcome) + as.factor(treatment), family = poisson())
 clean_names(m)
+
+# difference "clean_names()" and "find_variables()"
+library(lme4)
+m <- glmer(
+  cbind(incidence, size - incidence) ~ period + (1 | herd),
+  data = cbpp,
+  family = binomial
+)
+
+clean_names(m)
+find_variables(m)
+find_variables(m, flatten = TRUE)
+
 }
