+++ tests/testthat/test-blmer.R
@@ -104,7 +104,7 @@ if (require("testthat") && require("insight") && require("blme")) {
     expect_identical(find_terms(m1), list(
       response = "Reaction",
       conditional = "Days",
-      random = "Subject"
+      random = c("Days", "Subject")
     ))
     expect_identical(find_terms(m1, flatten = TRUE), c("Reaction", "Days", "Subject"))
     expect_identical(find_terms(m2), list(
@@ -113,6 +113,14 @@ if (require("testthat") && require("insight") && require("blme")) {
       random = c("mysubgrp", "mygrp", "Subject")
     ))
     expect_identical(find_terms(m2, flatten = TRUE), c("Reaction", "Days", "mysubgrp", "mygrp", "Subject"))
+  })
+
+  test_that("find_variables", {
+    expect_identical(find_variables(m1), list(
+      response = "Reaction",
+      conditional = "Days",
+      random = "Subject"
+    ))
   })
 
   test_that("get_response", {
