+++ tests/testthat/test-gamm4.R
@@ -50,8 +50,13 @@ if (require("testthat") && require("insight") && require("gamm4")) {
   })
 
   test_that("find_terms", {
-    expect_equal(find_terms(m1), list(response = "y", conditional = c("x0", "x1", "x2")))
-    expect_equal(find_terms(m1, flatten = TRUE), c("y", "x0", "x1", "x2"))
+    expect_equal(find_terms(m1), list(response = "y", conditional = c("s(x0)", "s(x1)", "s(x2)")))
+    expect_equal(find_terms(m1, flatten = TRUE), c("y", "s(x0)", "s(x1)", "s(x2)"))
+  })
+
+  test_that("find_variables", {
+    expect_equal(find_variables(m1), list(response = "y", conditional = c("x0", "x1", "x2")))
+    expect_equal(find_variables(m1, flatten = TRUE), c("y", "x0", "x1", "x2"))
   })
 
   test_that("n_obs", {
