+++ tests/testthat/test-rlmer.R
@@ -110,7 +110,7 @@ if (.runThisTest) {
       expect_identical(find_terms(m1), list(
         response = "Reaction",
         conditional = "Days",
-        random = "Subject"
+        random = c("Days", "Subject")
       ))
       expect_identical(find_terms(m1, flatten = TRUE), c("Reaction", "Days", "Subject"))
       expect_identical(find_terms(m2), list(
@@ -119,6 +119,14 @@ if (.runThisTest) {
         random = c("mysubgrp", "mygrp", "Subject")
       ))
       expect_identical(find_terms(m2, flatten = TRUE), c("Reaction", "Days", "mysubgrp", "mygrp", "Subject"))
+    })
+
+    test_that("find_variables", {
+      expect_identical(find_variables(m1), list(
+        response = "Reaction",
+        conditional = "Days",
+        random = "Subject"
+      ))
     })
 
     test_that("get_response", {
