### 2019-06-20 eb11259
fix tests
