+++ tests/testthat/test-Gam2.R
@@ -47,7 +47,7 @@ if (require("testthat") && require("insight") && require("gam")) {
 
   test_that("find_variables", {
     expect_equal(find_variables(m1), list(response = "Kyphosis", conditional = c("Age", "Number")))
-    expect_equal(find_terms(m1, flatten = TRUE), c("Kyphosis", "Age", "Number"))
+    expect_equal(find_variables(m1, flatten = TRUE), c("Kyphosis", "Age", "Number"))
   })
 
   test_that("n_obs", {
