+++ tests/testthat/test-mvrstanarm.R
@@ -47,9 +47,9 @@ if (.runThisTest && Sys.getenv("USER") != "travis") {
       expect_equal(colnames(get_response(m1)), c("logBili", "albumin"))
     })
 
-    test_that("find_terms", {
+    test_that("find_variables", {
       expect_identical(
-        find_terms(m1),
+        find_variables(m1),
         list(
           response = c(y1 = "logBili", y2 = "albumin"),
           y1 = list(conditional = "year", random = "id"),
@@ -57,16 +57,30 @@ if (.runThisTest && Sys.getenv("USER") != "travis") {
         )
       )
       expect_identical(
-        find_terms(m1, flatten = TRUE),
+        find_variables(m1, flatten = TRUE),
         c("logBili", "albumin", "year", "id", "sex")
       )
       expect_identical(
-        find_terms(m1, effects = "random"),
+        find_variables(m1, effects = "random"),
         list(
           response = c(y1 = "logBili", y2 = "albumin"),
           y1 = list(random = "id"),
           y2 = list(random = "id")
         )
+      )
+    })
+
+    test_that("find_terms", {
+      expect_identical(
+        find_terms(m1),
+        list(
+          y1 = list(response = "logBili", conditional = "year", random = "id"),
+          y2 = list(response = "albumin", conditional = c("sex", "year"), random = c("year", "id"))
+        )
+      )
+      expect_identical(
+        find_terms(m1, flatten = TRUE),
+        c("logBili", "year", "id", "albumin", "sex")
       )
     })
 
@@ -157,139 +171,798 @@ if (.runThisTest && Sys.getenv("USER") != "travis") {
     test_that("clean_parameters", {
       expect_identical(
         clean_parameters(m1),
-        structure(list(parameter = c("(Intercept)", "year", "sigma",
-        "(Intercept)", "sexf", "year", "sigma", "b[(Intercept) id:1]",
-        "b[(Intercept) id:2]", "b[(Intercept) id:3]", "b[(Intercept) id:4]",
-        "b[(Intercept) id:5]", "b[(Intercept) id:6]", "b[(Intercept) id:7]",
-        "b[(Intercept) id:8]", "b[(Intercept) id:9]", "b[(Intercept) id:10]",
-        "b[(Intercept) id:11]", "b[(Intercept) id:12]", "b[(Intercept) id:13]",
-        "b[(Intercept) id:14]", "b[(Intercept) id:15]", "b[(Intercept) id:16]",
-        "b[(Intercept) id:17]", "b[(Intercept) id:18]", "b[(Intercept) id:19]",
-        "b[(Intercept) id:20]", "b[(Intercept) id:21]", "b[(Intercept) id:22]",
-        "b[(Intercept) id:23]", "b[(Intercept) id:24]", "b[(Intercept) id:25]",
-        "b[(Intercept) id:26]", "b[(Intercept) id:27]", "b[(Intercept) id:28]",
-        "b[(Intercept) id:29]", "b[(Intercept) id:30]", "b[(Intercept) id:31]",
-        "b[(Intercept) id:32]", "b[(Intercept) id:33]", "b[(Intercept) id:34]",
-        "b[(Intercept) id:35]", "b[(Intercept) id:36]", "b[(Intercept) id:37]",
-        "b[(Intercept) id:38]", "b[(Intercept) id:39]", "b[(Intercept) id:40]",
-        "b[(Intercept) id:1]", "b[year id:1]", "b[(Intercept) id:2]",
-        "b[year id:2]", "b[(Intercept) id:3]", "b[year id:3]", "b[(Intercept) id:4]",
[…]
