+++ man/all_models_equal.Rd
@@ -32,9 +32,9 @@ m2 <- lm(mpg ~ wt + cyl, data = mtcars)
 m3 <- lmer(Reaction ~ Days + (1 | Subject), data = sleepstudy)
 m4 <- glm(formula = vs ~ wt, family = binomial(), data = mtcars)
 
-all_models_equal(m1, m2)
-all_models_equal(m1, m2, m3)
-all_models_equal(m1, m4, m2, m3, verbose = TRUE)
-all_models_equal(m1, m4, mtcars, m2, m3, verbose = TRUE)
+all_models_same_class(m1, m2)
+all_models_same_class(m1, m2, m3)
+all_models_same_class(m1, m4, m2, m3, verbose = TRUE)
+all_models_same_class(m1, m4, mtcars, m2, m3, verbose = TRUE)
 
 }
