### 2019-06-21 e321eeb
draft examples for paper
