### 2019-06-24 #1412 @strengejacke
Hi @alexpghayes

we have revised the paper, package and docs according to your last comments. Let me provide a point-by-point statement again to report the changes we made.

> Rename find_weights() -> weights_name() and get_weights() -> weights_values()

See our discussion above, for now we decided to stick to the function names, but we're thinking about introducing aliases where it make sense and doesn't obscure the documentation.

> I appreciated the new definitions of the various components of a model. Instead of presenting this material all at once, I recommend you group the definitions by components that are easy to confuse and set these groups visually apart.

We revised the paper and the docs by restructuring this section and changing the figures used to illustrate the different components. You find the changes in the paper [here](https://github.com/openjournals/joss-papers/blob/joss.01412/joss.01412/10.21105.joss.01412.pdf) and for the online-documentation [here](https://easystats.github.io/insight/articles/insight.html). Furthermore, we added these revisions to the [readme-page](https://github.com/easystats/insight/blob/master/README.md) and [documentation-homepage](https://easystats.github.io/insight/), however, w/o figures to keep the reading time for readme-page low.

> Note that you have chosen to define terms and variables in a non-standard way for the R community. In an R model formula an entire chunk such as I(x^2) is referred to as a term, rather than a variable. I recommend following this definition, which amounts to flipping the current usage of term and variable.

We have flipped the functions, so the function names are now in line with the common definitions inside (and probably outside) the R community, i.e. `find_terms()` is now named `find_variables()` and vice versa. All tests and docs were adapted accordingly, and so was the paper.

> Finally, I recommend grouping get_*() and find_*() together on the pkgdown site in the Reference section. That is, you want get_parameters() and find_parameters() (or parameter_names() and parameter_values()) to appear next to each other.

Thanks for the hint to this option (which we weren't aware of, but now also have applied to other package documentations). We have revised the docs accordingly and set groups visually apart here:
https://easystats.github.io/insight/reference/index.html

> It makes me nervous that the example at https://easystats.github.io/insight/reference/clean_parameters.html warns about overwriting quosure methods. Can you comment on why this is happening and/or necessary?

See our discussion above.

> When is clean_names() different from find_terms()? I would include examples so that users can tell when to use which function.

We have added more details and examples to the docs, making the differences between the two functions clearer: https://easystats.github.io/insight/reference/clean_names.html

> I appreciate that you've linked to packages that use insight. However, this is not especially useful to users -- instead of demonstrating how to solve a specific problem with insight, the user must now go read an entire package on their own, find a usage of insight, and infer the problem that is being solved. I would construct explicit, self-contained examples showing idiomatic usage. If these are inspired by the linked packages, fantastic, but merely linking to another code base isn't that helpful.

We have added another short section on practical use cases to the paper as well as to the readme-file and the documentation-homepage (https://easystats.github.io/insight/). These contain short code-snippets that should demonstrate the major aim of the package, its universal interface to access model information and how it helps users and package developers to reduce the hassle with different model objects.

> I recommend keeping the original paper title insight: Easy Access to Model Information for Various Model Objects rather than insight: A Unified Syntax for Accessing Model Information for Various Model Objects in R. Insight provides a unified interface, but does not introduce any new syntax, which is a language level feature. Domain specific languages such as dplyr and data.table introduce new syntax (i.e. := and !!) via metaprogramming tricks, but insight does not and makes use of standard evaluation and syntax.

We have renamed the paper title according to your suggestion and it now reads: _insight: A Unified Interface to Access Information from Model Objects in R_

> I would rename all_models_equal() to all_models_have_same_class(). The current name makes it seem like the all_models_equal() is also checking whether the parameter values match up.

We have renamed that functions, resp. added an alias to softly deprecate the old name.

> Instead of providing a new generic n_obs() I would provide methods for the existing generic nobs().

We have added a `nobs()` method for those model object that do not yet provide such a method, to not overwrite existing S3-methods.

> I am about to respond with additional thoughts on find_algorithm() in easystats/insight#38.

Thanks! We would like to keep this issue open until we have a clearer vision of how to improve this function, if you agree.

best
Daniel, on behalf of the co-authors
