### 2019-06-24 #1412 @alexpghayes
This looks great! @karthik I recommend publication.
