+++ R/n_obs.R
@@ -64,8 +64,8 @@ n_obs.gamm <- function(x, ...) {
 }
 
 #' @export
-nobs.gamm <- function(x, ...) {
-  x <- x$gam
+nobs.gamm <- function(object, ...) {
+  x <- object$gam
   class(x) <- c(class(x), c("glm", "lm"))
   n_obs(x, ...)
 }
@@ -78,7 +78,9 @@ n_obs.lmRob <- function(x, ...) {
 }
 
 #' @export
-nobs.lmRob <- n_obs.lmRob
+nobs.lmRob <- function(object, ...) {
+  n_obs(object, ...)
+}
 
 
 
@@ -94,7 +96,9 @@ n_obs.biglm <- function(x, ...) {
 }
 
 #' @export
-nobs.biglm <- n_obs.biglm
+nobs.biglm <- function(object, ...) {
+  n_obs(object, ...)
+}
 
 
 
@@ -104,7 +108,9 @@ n_obs.bigglm <- function(x, ...) {
 }
 
 #' @export
-nobs.bigglm <- n_obs.bigglm
+nobs.bigglm <- function(object, ...) {
+  n_obs(object, ...)
+}
 
 
 
@@ -114,7 +120,9 @@ n_obs.gbm <- function(x, ...) {
 }
 
 #' @export
-nobs.gbm <- n_obs.gbm
+nobs.gbm <- function(object, ...) {
+  n_obs(object, ...)
+}
 
 
 
@@ -124,7 +132,9 @@ n_obs.glimML <- function(x, ...) {
 }
 
 #' @export
-nobs.glimML <- n_obs.glimML
+nobs.glimML <- function(object, ...) {
+  n_obs(object, ...)
+}
 
 
 
@@ -134,7 +144,9 @@ n_obs.glmRob <- function(x, ...) {
 }
 
 #' @export
-nobs.glmRob <- n_obs.glmRob
+nobs.glmRob <- function(object, ...) {
+  n_obs(object, ...)
+}
 
 
 
@@ -144,7 +156,9 @@ n_obs.gmnl <- function(x, ...) {
 }
 
 #' @export
-nobs.gmnl <- n_obs.gmnl
+nobs.gmnl <- function(object, ...) {
+  n_obs(object, ...)
+}
 
 
 
@@ -155,7 +169,9 @@ n_obs.multinom <- function(x, ...) {
 }
 
 #' @export
-nobs.multinom <- n_obs.multinom
+nobs.multinom <- function(object, ...) {
+  n_obs(object, ...)
+}
 
 
 
@@ -165,7 +181,9 @@ n_obs.rq <- function(x, ...) {
 }
 
 #' @export
-nobs.rq <- n_obs.rq
+nobs.rq <- function(object, ...) {
+  n_obs(object, ...)
+}
 
 
 
@@ -175,7 +193,9 @@ n_obs.BBreg <- function(x, ...) {
 }
 
 #' @export
-nobs.BBreg <- n_obs.BBreg
+nobs.BBreg <- function(object, ...) {
+  n_obs(object, ...)
+}
 
 
 
@@ -186,7 +206,9 @@ n_obs.BBmm <- function(x, ...) {
 }
 
 #' @export
-nobs.BBmm <- n_obs.BBmm
+nobs.BBmm <- function(object, ...) {
+  n_obs(object, ...)
+}
 
 
 
@@ -200,7 +222,9 @@ n_obs.crq <- function(x, ...) {
 }
 
 #' @export
-nobs.crq <- n_obs.crq
+nobs.crq <- function(object, ...) {
+  n_obs(object, ...)
+}
 
 
 
[…]
