+++ NAMESPACE
@@ -379,6 +379,7 @@ S3method(n_obs,multinom)
 S3method(n_obs,rq)
 S3method(n_obs,stanmvreg)
 S3method(n_obs,survfit)
+S3method(n_obs,survreg)
 S3method(n_obs,svyolr)
 S3method(n_obs,wbm)
 S3method(n_obs,zeroinfl)
@@ -405,6 +406,7 @@ S3method(nobs,multinom)
 S3method(nobs,rq)
 S3method(nobs,stanmvreg)
 S3method(nobs,survfit)
+S3method(nobs,survreg)
 S3method(nobs,zeroinfl)
 S3method(nobs,zerotrunc)
 export(all_models_equal)
