+++ R/clean_parameters.R
@@ -33,6 +33,7 @@
 #' or for specific terms like smooth- or spline-terms).
 #'
 #' @examples
+#' library(brms)
 #' model <- download_model("brms_zi_2")
 #' clean_parameters(model)
 #' @export
