+++ man/clean_parameters.Rd
@@ -43,6 +43,7 @@ contains "human readable" parameter names (which are mostly identical to
 or for specific terms like smooth- or spline-terms).
 }
 \examples{
+library(brms)
 model <- download_model("brms_zi_2")
 clean_parameters(model)
 }
