### 2019-06-25 2407c4c
A few more grammatical catches/fixes
