### 2019-06-25 #1412 @karthik
@alexpghayes Thanks so much for such a thorough review. Your review has improved this package far beyond the checklist we have here. 🙏 💯

@strengejacke and co-authors: The paper in the current form is a little too long for JOSS. JOSS papers are not meant to replace documentation, and should really only provide a high level overview of the package and highlight a few applications. It should not read like a vignette. I'd recommend taking out

Example use cases in R
Examples of use cases in R packages (and linking to both sections in your package down docs). Once you make these edits I'm happy to proceed with the remaining steps.
