### 2019-06-25 #1412 @strengejacke
Ok, I revised the paper accordingly and just checked the PDF proof. Next step would be creating an archive with a DOI, right?
