### 2019-06-25 eb546f1
update bib-file
