### 2019-06-25 #1412 @strengejacke
Some references have been deleted during the removal of the paragraphs you suggested, so the reference list is shorter now. Maybe it was the bib-file, which still contained the old entries? What about now?
