### 2019-06-25 #1412 @strengejacke
Ok, I reduced the parapgraphs into one sentence, inlcuding the references (I added them back now), and referring to the website. So we have removed the long sections, but still have all references in the paper as before. The new DOI for this archive is: 10.5281/zenodo.3256639

I'll generate a PDF proof to check.
