### 2019-06-25 #1412 @kyleniemeyer
Congrats on your article's acceptance in JOSS @strengejacke! Thanks to @alexpghayes for reviewing and @karthik for editing.
