### 2019-06-25 #1412 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
